Census by nodes or clusters

These functions include ways to take a census of the positions of nodes in a network. These include a triad census based on the triad profile of nodes, but also a tie census based on the particular tie partners of nodes. Included also are group census functions for summarising the profiles of clusters of nodes in a network.

node_tie_census(object)

node_triad_census(object)

group_tie_census(object, clusters, decimals = 2)

group_triad_census(object, clusters, decimals = 2)

Arguments

object
a migraph-consistent object

clusters
a vector of cluster assignment

decimals
number of decimal points to round to

Examples

task_eg <- to_named(to_uniplex(ison_m182, "task_tie"))
(tie_cen <- node_tie_census(task_eg))

#>         fromCarole fromGayle fromCiara fromRamona fromHarold fromCharley
#> Carole        0.00      0.00      0.00       0.00       0.30        0.00
#> Gayle         0.00      0.00      0.00       0.00       0.00        0.00
#> Ciara         0.00      0.00      0.00       0.00       0.15        0.75
#> Ramona        0.00      0.00      0.00       0.00       0.00        0.00
#> Harold        0.30      0.00      0.15       0.00       0.00        0.45
#> Charley       0.00      0.00      0.60       0.00       0.75        0.00
#> Suzanne       0.00      0.75      0.00       0.00       0.00        0.00
#> Derrick       0.00      0.30      0.00       0.00       0.00        0.00
#> Kiara         0.30      0.00      0.00       0.00       0.00        0.00
#> Tamika        0.30      0.00      0.00       0.15       0.00        0.15
#> Kent          0.00      0.00      0.30       0.00       2.10        0.30
#> Rachel        0.15      0.00      0.00       0.00       0.00        0.00
#> Braden        0.00      0.90      0.00       0.00       0.15        0.00
#> Isla          0.00      0.15      0.00       0.00       0.00        0.00
#> Chris         0.15      0.15      0.00       0.00       0.00        0.00
#> Bettie        5.10      4.80      0.30       0.90       1.80        0.90
#>         fromSuzanne fromDerrick fromKiara fromTamika fromKent fromRachel
#> Carole         0.00        0.00      0.15       0.15     0.00       0.00
#> Gayle          1.20        0.45      0.00       0.00     0.00       0.00
#> Ciara          0.00        0.00      0.00       0.00     0.15       0.00
#> Ramona         0.00        0.00      0.00       0.15     0.00       0.00
#> Harold         0.00        0.00      0.00       0.00     1.80       0.00
#> Charley        0.00        0.00      0.00       0.30     0.45       0.00
#> Suzanne        0.00        0.30      0.00       0.00     0.00       0.00
#> Derrick        0.15        0.00      0.00       0.00     0.00       0.00
#> Kiara          0.00        0.00      0.00       0.45     0.00       2.40
#> Tamika         0.00        0.00      0.30       0.00     0.00       0.45
#> Kent           0.00        0.00      0.00       0.00     0.00       0.00
#> Rachel         0.00        0.00      2.70       0.75     0.00       0.00
#> Braden         0.30        0.75      0.00       0.00     0.00       0.00
#> Isla           0.00        0.15      0.00       0.00     0.00       0.00
#> Chris          0.00        0.00      0.15       0.30     0.00       0.75
#> Bettie         2.85        2.40      2.40       5.40     1.05       4.50
#>         fromBraden fromIsla fromChris fromBettie toCarole toGayle toCiara
#> Carole        0.00     0.00      0.15       4.65     0.00    0.00    0.00
#> Gayle         1.05     0.30      0.00       4.80     0.00    0.00    0.00
#> Ciara         0.00     0.00      0.00       0.75     0.00    0.00    0.00
#> Ramona        0.00     0.00      0.00       1.05     0.00    0.00    0.00
#> Harold        0.00     0.15      0.00       2.40     0.30    0.00    0.15
#> Charley       0.00     0.15      0.00       1.20     0.00    0.00    0.75
#> Suzanne       0.30     0.15      0.00       2.40     0.00    1.20    0.00
#> Derrick       0.60     0.30      0.00       1.95     0.00    0.45    0.00
#> Kiara         0.00     0.00      0.15       2.85     0.15    0.00    0.00
#> Tamika        0.00     0.00      0.30       5.40     0.15    0.00    0.00
#> Kent          0.00     0.00      0.00       1.20     0.00    0.00    0.15
#> Rachel        0.00     0.00      0.45       5.25     0.00    0.00    0.00
#> Braden        0.00     0.00      0.00       3.90     0.00    1.05    0.00
#> Isla          0.00     0.00      0.00       3.90     0.00    0.30    0.00
#> Chris         0.00     0.00      0.00       5.55     0.15    0.00    0.00
#> Bettie        4.50     3.15      6.60       0.00     4.65    4.80    0.75
#>         toRamona toHarold toCharley toSuzanne toDerrick toKiara toTamika toKent
#> Carole      0.00     0.30      0.00      0.00      0.00    0.30     0.30    0.0
#> Gayle       0.00     0.00      0.00      0.75      0.30    0.00     0.00    0.0
#> Ciara       0.00     0.15      0.60      0.00      0.00    0.00     0.00    0.3
#> Ramona      0.00     0.00      0.00      0.00      0.00    0.00     0.15    0.0
#> Harold      0.00     0.00      0.75      0.00      0.00    0.00     0.00    2.1
#> Charley     0.00     0.45      0.00      0.00      0.00    0.00     0.15    0.3
#> Suzanne     0.00     0.00      0.00      0.00      0.15    0.00     0.00    0.0
#> Derrick     0.00     0.00      0.00      0.30      0.00    0.00     0.00    0.0
#> Kiara       0.00     0.00      0.00      0.00      0.00    0.00     0.30    0.0
#> Tamika      0.15     0.00      0.30      0.00      0.00    0.45     0.00    0.0
#> Kent        0.00     1.80      0.45      0.00      0.00    0.00     0.00    0.0
#> Rachel      0.00     0.00      0.00      0.00      0.00    2.40     0.45    0.0
#> Braden      0.00     0.00      0.00      0.30      0.60    0.00     0.00    0.0
#> Isla        0.00     0.15      0.15      0.15      0.30    0.00     0.00    0.0
#> Chris       0.00     0.00      0.00      0.00      0.00    0.15     0.30    0.0
#> Bettie      1.05     2.40      1.20      2.40      1.95    2.85     5.40    1.2
#>         toRachel toBraden toIsla toChris toBettie
#> Carole      0.15     0.00   0.00    0.15     5.10
#> Gayle       0.00     0.90   0.15    0.15     4.80
#> Ciara       0.00     0.00   0.00    0.00     0.30
#> Ramona      0.00     0.00   0.00    0.00     0.90
#> Harold      0.00     0.15   0.00    0.00     1.80
#> Charley     0.00     0.00   0.00    0.00     0.90
#> Suzanne     0.00     0.30   0.00    0.00     2.85
#> Derrick     0.00     0.75   0.15    0.00     2.40
#> Kiara       2.70     0.00   0.00    0.15     2.40
#> Tamika      0.75     0.00   0.00    0.30     5.40
#> Kent        0.00     0.00   0.00    0.00     1.05
#> Rachel      0.00     0.00   0.00    0.75     4.50
#> Braden      0.00     0.00   0.00    0.00     4.50
#> Isla        0.00     0.00   0.00    0.00     3.15
#> Chris       0.45     0.00   0.00    0.00     6.60
#> Bettie      5.25     3.90   3.90    5.55     0.00

(triad_cen <- node_triad_census(task_eg))

#>         003 012 102 021D 021U 021C 111D 111U 030T 030C 201 120D 120U 120C 210
#> Carole   36   9  45    0    0    0    0    1    0    0   3    0    0    0   4
#> Gayle    45  10  38    0    0    0    1    1    0    0   1    0    0    0   3
#> Ciara    55   0  44    0    0    0    0    0    0    0   0    0    0    0   0
#> Ramona  102   0   3    0    0    0    0    0    0    0   0    0    0    0   0
#> Harold   55   0  37    0    0    0    2    1    0    0   4    0    0    0   0
#> Charley  65  19  10    0    0    0    5    0    0    0   3    1    0    0   2
#> Suzanne  91   1  13    0    0    0    0    0    0    0   0    0    0    0   0
#> Derrick  91   1  13    0    0    0    0    0    0    0   0    0    0    0   0
#> Kiara    90   1  14    0    0    0    0    0    0    0   0    0    0    0   0
#> Tamika   86   1  18    0    0    0    0    0    0    0   0    0    0    0   0
#> Kent     95   0  10    0    0    0    0    0    0    0   0    0    0    0   0
#> Rachel   90   1  14    0    0    0    0    0    0    0   0    0    0    0   0
#> Braden   93   1  11    0    0    0    0    0    0    0   0    0    0    0   0
#> Isla     90   3  12    0    0    0    0    0    0    0   0    0    0    0   0
#> Chris    88   2  15    0    0    0    0    0    0    0   0    0    0    0   0
#> Bettie    0   0   0    0    0    0    0    0    0    0  73    0    0    0   6
#>         300
#> Carole    7
#> Gayle     6
#> Ciara     6
#> Ramona    0
#> Harold    6
#> Charley   0
#> Suzanne   0
#> Derrick   0
#> Kiara     0
#> Tamika    0
#> Kent      0
#> Rachel    0
#> Braden    0
#> Isla      0
#> Chris     0
#> Bettie   26

group_tie_census(task_eg, cutree(cluster_structural_equivalence(task_eg), 4))

#>         fromCarole fromGayle fromCiara fromRamona fromHarold fromCharley
#> Block 1       0.08       0.2      0.00       0.01       0.04        0.01
#> Block 2       0.15       0.0      0.07       0.00       0.07        0.60
#> Block 3       0.00       0.0      0.45       0.00       1.43        0.15
#> Block 4       5.10       4.8      0.30       0.90       1.80        0.90
#>         fromSuzanne fromDerrick fromKiara fromTamika fromKent fromRachel
#> Block 1        0.15        0.15       0.3       0.16     0.00       0.33
#> Block 2        0.00        0.00       0.0       0.00     0.98       0.00
#> Block 3        0.00        0.00       0.0       0.15     0.22       0.00
#> Block 4        2.85        2.40       2.4       5.40     1.05       4.50
#>         fromBraden fromIsla fromChris fromBettie toCarole toGayle toCiara
#> Block 1       0.18     0.07       0.1       3.79     0.04    0.27    0.00
#> Block 2       0.00     0.07       0.0       1.57     0.15    0.00    0.07
#> Block 3       0.00     0.07       0.0       1.20     0.00    0.00    0.45
#> Block 4       4.50     3.15       6.6       0.00     4.65    4.80    0.75
#>         toRamona toHarold toCharley toSuzanne toDerrick toKiara toTamika toKent
#> Block 1     0.01     0.04      0.04      0.14      0.12    0.30     0.14   0.00
#> Block 2     0.00     0.07      0.68      0.00      0.00    0.00     0.00   1.20
#> Block 3     0.00     1.12      0.22      0.00      0.00    0.00     0.07   0.15
#> Block 4     1.05     2.40      1.20      2.40      1.95    2.85     5.40   1.20
#>         toRachel toBraden toIsla toChris toBettie
#> Block 1     0.37     0.18   0.03    0.14     3.87
#> Block 2     0.00     0.07   0.00    0.00     1.05
#> Block 3     0.00     0.00   0.00    0.00     0.98
#> Block 4     5.25     3.90   3.90    5.55     0.00

group_triad_census(task_eg, cutree(cluster_regular_equivalence(task_eg), 4))

#>           003  012   102 021D 021U 021C 111D 111U 030T 030C   201 120D 120U
#> Block 1 36.00 9.00 45.00    0    0    0 0.00 1.00    0    0  3.00 0.00    0
#> Block 2 51.67 3.33 39.67    0    0    0 1.00 0.67    0    0  1.67 0.00    0
#> Block 3 89.18 2.73 12.09    0    0    0 0.45 0.00    0    0  0.27 0.09    0
#> Block 4  0.00 0.00  0.00    0    0    0 0.00 0.00    0    0 73.00 0.00    0
#>         120C  210 300
#> Block 1    0 4.00   7
#> Block 2    0 1.00   6
#> Block 3    0 0.18   0
#> Block 4    0 6.00  26