Measures of network centralisation

Measures of network centralisation

graph_degree(object, normalized = TRUE, direction = c("all", "out", "in"))

graph_closeness(object, normalized = TRUE, direction = c("all", "out", "in"))

graph_betweenness(object, normalized = TRUE, direction = c("all", "out", "in"))

graph_eigenvector(object, normalized = TRUE)

Arguments

object

An object of a migraph-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

normalized

Logical scalar, whether the centrality scores are normalized. Different denominators are used depending on whether the object is one-mode or two-mode, the type of centrality, and other arguments.

direction

Character string, “out” bases the measure on outgoing ties, “in” on incoming ties, and "all" on either/the sum of the two. For two-mode networks, "all" uses as numerator the sum of differences between the maximum centrality score for the mode against all other centrality scores in the network, whereas "in" uses as numerator the sum of differences between the maximum centrality score for the mode against only the centrality scores of the other nodes in that mode.

Functions

graph_degree: Calculate the degree centralization for a graph

graph_closeness: Calculate the closeness centralization for a graph

graph_betweenness: Calculate the betweenness centralization for a graph

graph_eigenvector: Calculate the eigenvector centralization for a graph

See also

Examples

graph_degree(ison_southern_women, direction = "in")
#> Mode 1 Mode 2 
#>  0.231  0.466 
graph_closeness(ison_southern_women, direction = "in")
#> Mode 1 Mode 2 
#>  0.214  0.528 
graph_betweenness(ison_southern_women, direction = "in")
#> Mode 1 Mode 2 
#> 0.0668 0.1982 
graph_eigenvector(mpn_elite_mex)
#> [1] 0.63
graph_eigenvector(ison_southern_women)
#> Mode 1 Mode 2 
#> 0.0849 0.2630