Centrality for one- and two-mode networks

These functions calculate common centrality measures for both one- and two-mode networks. They accept as objects matrices and igraph graphs, and can be used within a tidygraph workflow. Importantly, these functions also offer correct normalization for two-mode networks.

node_degree(
  object,
  weights = NULL,
  mode = "out",
  loops = TRUE,
  normalized = FALSE
)

node_closeness(
  object,
  weights = NULL,
  mode = "out",
  normalized = FALSE,
  cutoff = NULL
)

node_betweenness(
  object,
  weights = NULL,
  directed = TRUE,
  cutoff = NULL,
  nobigint = TRUE,
  normalized = FALSE
)

node_eigenvector(
  object,
  weights = NULL,
  directed = FALSE,
  scale = FALSE,
  normalized = FALSE
)

Arguments

object

Either an igraph graph object or a matrix.

weights

The weight of the edges to use for the calculation. Will be evaluated in the context of the edge data.

mode

How should edges be followed (in or out). By default, outdegree of the node is calculated. Ignored for undirected graphs.

loops

Should loops be included in the calculation

normalized

For one-mode networks, should Borgatti and Everett normalization be applied?

cutoff

maximum path length to use during calculations

directed

Should direction of edges be used for the calculations

nobigint

Should big integers be avoided during calculations

scale

Should the scores be scaled to range between 0 and 1?

Value

Depending on how and what kind of an object is passed to the function, the function will return a tidygraph object where the nodes have been updated A numeric vector giving the betweenness centrality measure of each node. A numeric vector giving the eigenvector centrality measure of each node.

References

Borgatti, Stephen P., and Martin G. Everett (1997). "Network analysis of 2-mode data." Social Networks 19(3): 243-270.

Faust, Katherine (1997). "Centrality in affiliation networks." Social Networks 19(2): 157-191.

See also

Examples

node_degree(mpn_elite_mex)
#>     Trevino      Madero    Carranza     Aguilar     Obregon      Calles 
#> 3           6           8           6           6           6           
#>             
#> + 29 others 
node_degree(ison_southern_women)
#>      EVELYN       LAURA     THERESA      BRENDA   CHARLOTTE     FRANCES 
#> 8           7           8           7           4           4           
#>                      E9         E10         E11         E12         E13 
#> + 20 others 12           6           4           7           4          
#>         E14 
#>  4          
node_closeness(mpn_elite_mex)
#>     Trevino      Madero    Carranza     Aguilar     Obregon      Calles 
#> 0.0118      0.0119      0.0137      0.0145      0.0128      0.0135      
#>             
#> + 29 others 
node_closeness(ison_southern_women)
#>      EVELYN       LAURA     THERESA      BRENDA   CHARLOTTE     FRANCES 
#> 0.0167      0.0152      0.0167      0.0152      0.0125      0.0139      
#>                      E9         E10         E11         E12         E13 
#> + 20 others 0.0179      0.0128      0.0122      0.0132      0.0122      
#>         E14 
#> 0.0122      
node_betweenness(mpn_elite_mex)
#>     Trevino      Madero    Carranza     Aguilar     Obregon      Calles 
#>  2.83        4.59       17.36       11.56        4.52       13.96       
#>             
#> + 29 others 
node_betweenness(ison_southern_women)
#>      EVELYN       LAURA     THERESA      BRENDA   CHARLOTTE     FRANCES 
#> 42.76       22.86       38.74       22.01        4.73        4.75       
#>                      E9         E10         E11         E12         E13 
#> + 20 others 96.23        6.82        9.02       10.24        1.89       
#>         E14 
#>  1.89       
node_eigenvector(mpn_elite_mex)
#>     Trevino      Madero    Carranza     Aguilar     Obregon      Calles 
#> 0.0571      0.0771      0.1176      0.1204      0.0918      0.0977      
#>             
#> + 29 others 
node_eigenvector(ison_southern_women)
#>      EVELYN       LAURA     THERESA      BRENDA   CHARLOTTE     FRANCES 
#> 0.299       0.281       0.334       0.284       0.161       0.203       
#>                      E9         E10         E11         E12         E13 
#> + 20 others 0.357       0.228       0.112       0.255       0.177       
#>         E14 
#> 0.177