Centralization for one- and two-mode networks

These functions measure the overall centralization for a network.

graph_degree(
  object,
  directed = c("all", "out", "in", "total"),
  normalized = TRUE
)

graph_closeness(
  object,
  directed = c("all", "out", "in", "total"),
  normalized = TRUE
)

graph_betweenness(
  object,
  directed = c("all", "out", "in", "total"),
  normalized = TRUE
)

Arguments

object
A matrix, igraph graph, or tidygraph object.

directed
Character string, “out” for out-degree, “in” for in-degree, and "all" or “total” for the sum of the two. For two-mode networks, "all" uses as numerator the sum of differences between the maximum centrality score for the mode against all other centrality scores in the network, whereas "in" uses as numerator the sum of differences between the maximum centrality score for the mode against only the centrality scores of the other nodes in that nodeset.

normalized
Logical scalar, whether the centrality scores are normalized. Different denominators are used depending on whether the object is one-mode or two-mode, the type of centrality, and other arguments.

Value

A single centralization score if the object was one-mode, and two centralization scores if the object was two-mode. In the case of a two-mode network, to return just the score for the first nodeset (rows), append `$nodes1` to the end of the function call or returned object. To return just the score for the second nodeset (cols), append `$nodes2` to the end of the function call or returned object.

References

Borgatti, Stephen P, and Daniel S Halgin. 2011. ``Analyzing Affiliation Networks." In The SAGE Handbook of Social Network Analysis, edited by John Scott and Peter J Carrington, 417–33. London, UK: Sage.

See also

Examples

graph_degree(southern_women, directed = "in")

#> $nodes1
#> [1] 0.2307692
#> 
#> $nodes2
#> [1] 0.4660633
#> 

graph_closeness(southern_women, directed = "in")

#> $nodes1
#> [1] 0.2135132
#> 
#> $nodes2
#> [1] 0.5284847
#> 

graph_betweenness(southern_women, directed = "in")

#> $nodes1
#> [1] 0.06679267
#> 
#> $nodes2
#> [1] 0.1982068
#>