Measures of network closure

These functions offer methods for summarising the closure in configurations in one-, two-, and three-mode networks.

network_reciprocity(.data, method = "default")

node_reciprocity(.data)

network_transitivity(.data)

node_transitivity(.data)

network_equivalency(.data)

network_congruency(.data, object2)

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

method

For reciprocity, either default or ratio. See ?igraph::reciprocity

object2

Optionally, a second (two-mode) matrix, igraph, or tidygraph

Details

For one-mode networks, shallow wrappers of igraph versions exist via network_reciprocity and network_transitivity.

For two-mode networks, network_equivalency calculates the proportion of three-paths in the network that are closed by fourth tie to establish a "shared four-cycle" structure.

For three-mode networks, network_congruency calculates the proportion of three-paths spanning two two-mode networks that are closed by a fourth tie to establish a "congruent four-cycle" structure.

Functions

network_reciprocity(): Calculate reciprocity in a (usually directed) network

node_reciprocity(): Calculate nodes' reciprocity

network_transitivity(): Calculate transitivity in a network

node_transitivity(): Calculate nodes' transitivity

network_equivalency(): Calculate equivalence or reinforcement in a (usually two-mode) network

network_congruency(): Calculate congruency across two two-mode networks

References

Robins, Garry L, and Malcolm Alexander. 2004. Small worlds among interlocking directors: Network structure and distance in bipartite graphs. Computational & Mathematical Organization Theory 10(1): 69–94. doi:10.1023/B:CMOT.0000032580.12184.c0 .

Knoke, David, Mario Diani, James Hollway, and Dimitris C Christopoulos. 2021. Multimodal Political Networks. Cambridge University Press. Cambridge University Press. doi:10.1017/9781108985000

See also

Examples

network_reciprocity(ison_southern_women)
#> [1] 1
node_reciprocity(to_unweighted(ison_networkers))
#>   `Lin Freeman` `Doug White` `Ev Rogers` `Richard Alba` `Phipps Arabie`
#> 1         0.935         0.75           1          0.944           0.286
#> # ... with 27 more values from this nodeset unprinted. Use `print(..., n = Inf)` to print all values.
network_transitivity(ison_adolescents)
#> [1] 0.45
node_transitivity(ison_adolescents)
#>   Betty   Sue Alice  Jane  Dale   Pam Carol  Tina
#> 1   NaN 0.333   0.5     1 0.667 0.333     0   NaN
network_equivalency(ison_southern_women)
#> [1] 0.487