Cohesion for one-, two-, and three- mode networks

These functions offer methods for summarising the cohesion in one-, two-, and three-mode networks.

graph_density(object)

graph_reciprocity(object, method = "default")

graph_transitivity(object)

graph_equivalency(object)

graph_congruency(object, object2)

Arguments

object

A one-mode or two-mode matrix, igraph, or tidygraph

method

For reciprocity, either default or ratio. See ?igraph::reciprocity

object2

Optionally, a second (two-mode) matrix, igraph, or tidygraph

Details

For one- and two-mode networks, graph_density summarises the ratio of ties to the number of possible ties.

For one-mode networks, shallow wrappers of igraph versions exist via graph_reciprocity and graph_transitivity.

For two-mode networks, graph_equivalency calculates the proportion of three-paths in the network that are closed by fourth tie to establish a "shared four-cycle" structure.

For three-mode networks, graph_congruency calculates the proportion of three-paths spanning the two two-mode networks that are closed by a fourth tie to establish a "congruent four-cycle" structure.

References

Robins, Garry L, and Malcolm Alexander. 2004. Small worlds among interlocking directors: Network structure and distance in bipartite graphs. Computational & Mathematical Organization Theory 10 (1): 69–94.

Knoke, David, Mario Diani, James Hollway, and Dimitris C Christopoulos. 2021. Multimodal Political Networks. Cambridge University Press. Cambridge University Press.

See also

Examples

graph_density(mpn_elite_mex)
#> [1] 0.1966387
graph_density(mpn_elite_usa_advice)
#> [1] 0.1642857
graph_reciprocity(ison_southern_women)
#> [1] 1
graph_transitivity(ison_southern_women)
#> [1] 0
graph_equivalency(ison_southern_women)
#> [1] 0.4871634