Component partitioning algorithms

These functions create a vector of nodes' memberships in components or degrees of coreness.

In graph theory, components, sometimes called connected components, are induced subgraphs from partitioning the nodes into disjoint sets. All nodes that are members of the same partition as i are reachable from i.

For directed networks, strongly connected components consist of subgraphs where there are paths in each direction between member nodes. Weakly connected components consist of subgraphs where there is a path in either direction between member nodes.

Coreness captures the maximal subgraphs in which each vertex has at least degree k, where k is also the order of the subgraph. As described in igraph::coreness, a node's coreness is k if it belongs to the k-core but not to the (k+1)-core.

node_components(.data)

node_weak_components(.data)

node_strong_components(.data)

node_coreness(.data)

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

Functions

node_components(): Returns nodes' component membership using edge direction where available.

node_weak_components(): Returns nodes' component membership ignoring edge direction.

node_strong_components(): Returns nodes' component membership based on edge direction.

node_coreness(): Returns k-cores

See also

Examples

node_components(mpn_bristol)
#> This graph was created by an old(er) igraph version.
#>   Call upgrade_graph() on it to use with the current igraph version
#>   For now we convert it on the fly...
#> 1 
#>    AddReAg, AfrInit, OrganicCoop, Wildlife, BACEN, BangladAss, Barnardos, BartonHill, BigIssue, BREAD, Stroke, CivicSoc, CyclingCamp, DialRide, Foyer, BHAM, InterFaith, JamiaMosque, MuslimCulSoc, BVEG, WestIndianParents, CareForum, CEED, DeafPeople, CentrSustEnergy, CityFarms, CND, CoopParty, COGB, RecyclNet, CSV, CycleWest, AsylumSeekers, DhekBhal, Disabled, DundryHill, EBYHousAss, EastonAsianWom, ECA, FoE, SevernRai, FullCircle, Greens, Greenleaf, Greenpeace, H&WTeenageParents, HartcliffeClubYoung, HartCommCampus, HHEAG, ISIS-MP, Kebele, KHAAS, KWDrugs, KWMedia, KWYouth, LawCentre, LGBForum, LivingEaston, Malago, Maternity, Mede, Mind, NextLink, NextStep, Nigerian, NSForum, CommonFuture, Oxfam, PakistaniWomen, Permaculture, PinkParents, PIPA, PrincessCarers, PriorYouthHousing, RailwayDevelop, RecConsort, SARI, Schumacher, SecondStep, SERA, SikhRes, SocialistParty, SOFA, Soil, SBCC, SPAdvice, StWerbergh, StopOldbury, Sustrans, Transform, UnityGroup, WECIL, Wellwomen, Womankind, WWF, CommunityFest, EastonFest, H&WCarnival, LA21, Respect, SortItFest, StPaulCarnival, WomensDay, AshtonQuarry, AsylumCampaign, Babymilk, Claimants, GlobalResist, Ikea, Jubilee2000, M32Reclaim, STAR 
#>    101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 111, 112, 113, 114, 115, 116, 117, 118, 119, 120, 121, 122, 123, 124, 125, 126, 127, 128, 129, 130, 133, 134, 135, 136, 137, 138, 139, 140, 141, 142, 143, 144, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 158, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169, 170, 171, 172, 173, 174, 175, 176, 177, 178, 179, 180, 181, 182, 183, 184, 185, 186, 187, 188, 189, 190, 191, 193, 194, 196, 197, 198, 199, 200, 201, 202, 203, 204, 205, 206, 207, 208, 209, 210, 211, 212, 213, 214, 215, 216, 217, 218, 219, 220, 222, 223, 224, 225, 226, 227, 228, 229, 230, 231, 232, 233, 234, 235, 236, 237, 238, 239, 240, 241, 242, 243, 244, 245, 246, 247, 248, 249, 250, 251, 252, 253
#>  2 
#>    LocalFood 
#>    131
#>  3 
#>    AvonBat 
#>    192
node_coreness(ison_adolescents)
#> 1 
#>    Betty, Carol, Tina
#>  2 
#>    Sue, Alice, Jane, Dale, Pam