Core-periphery clustering algorithms

These functions identify nodes belonging to (some level of) the core of a network:

node_core() assigns nodes to either the core or periphery.

node_coreness() assigns nodes to their level of k-coreness.

node_core(.data, method = c("degree", "eigenvector"))

node_coreness(.data)

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

method

Which method to use to identify cores and periphery. By default this is "degree", which relies on the heuristic that high degree nodes are more likely to be in the core. An alternative is "eigenvector", which instead begins with high eigenvector nodes. Other methods, such as a genetic algorithm, CONCOR, and Rombach-Porter, can be added if there is interest.

Core-periphery

This function is used to identify which nodes should belong to the core, and which to the periphery. It seeks to minimize the following quantity: $$Z(S_1) = \sum_{(i<j)\in S_1} \textbf{I}_{\{A_{ij}=0\}} + \sum_{(i<j)\notin S_1} \textbf{I}_{\{A_{ij}=1\}}$$ where nodes \(\{i,j,...,n\}\) are ordered in descending degree, \(A\) is the adjacency matrix, and the indicator function is 1 if the predicate is true or 0 otherwise. Note that minimising this quantity maximises density in the core block and minimises density in the periphery block; it ignores ties between these blocks.

References

Borgatti, Stephen P., & Everett, Martin G. 1999. Models of core /periphery structures. Social Networks, 21, 375–395. doi:10.1016/S0378-8733(99)00019-2

Lip, Sean Z. W. 2011. “A Fast Algorithm for the Discrete Core/Periphery Bipartitioning Problem.” doi:10.48550/arXiv.1102.5511

See also

Examples

#mpn_elite_usa_advice %>% as_tidygraph %>% 
#   mutate(corep = node_core(mpn_elite_usa_advice)) %>% 
#   autographr(node_color = "corep")
network_core(mpn_elite_usa_advice)
#> [1] 0.338
node_coreness(ison_adolescents)
#>   Betty   Sue Alice  Jane  Dale   Pam Carol  Tina
#> 1     1     2     2     2     2     2     1     1