Functions that have been renamed, superseded, or are no longer working

Generally these functions have been superseded or renamed. Upon using them, a message is provided directing the user to the new function. However, at this stage of package development, we generally clear older defunct functions at each minor release, and so you are strongly encouraged to use the new functions/names/syntax wherever possible and update your scripts accordingly.

test_gof(diff_model, diff_models)

network_reg(
  formula,
  .data,
  method = c("qap", "qapy"),
  times = 1000,
  strategy = "sequential",
  verbose = FALSE
)

Functions

test_gof(): Deprecated on 2024-06-16.

network_reg(): Deprecated on 2024-07-19.