Clustering for one-, two-, and three- mode networks

This function offers clustering methods for one-, two-, and three-mode networks.

graph_clustering(object, object2 = NULL)

Arguments

object
A one-mode or two-mode matrix, igraph, or tidygraph

object2
Optionally, a second (two-mode) matrix, igraph, or tidygraph

Details

For one-mode networks, the function serves as a shallow wrapper for `igraph::transitivity`, since global transitivity is a regular measure for clustering or local density in one-mode networks.

For two-mode networks, we calculate the proportion of three-paths in the network that are closed by fourth tie to establish a "shared four-cycle" structure.

For three-mode networks, we calculate the proportion of three-paths spanning the two two-mode networks that are closed by a fourth tie to establish a "congruent four-cycle" structure.

References

Robins, Garry L, and Malcolm Alexander. 2004. Small worlds among interlocking directors: Network structure and distance in bipartite graphs. Computational & Mathematical Organization Theory 10 (1): 69–94.

Knoke, David, Mario Diani, James Hollway, and Dimitris C Christopoulos. 2021. Multimodal Political Networks. Cambridge University Press. Cambridge University Press.

See also

Examples

graph_clustering(southern_women)

#> [1] 0.4871634