Graph theoretic dimensions of hierarchy

These functions, together with network_reciprocity(), are used jointly to measure how hierarchical a network is:

network_connectedness() measures the proportion of dyads in the network that are reachable to one another, or the degree to which network is a single component.

network_efficiency() measures the Krackhardt efficiency score.

network_upperbound() measures the Krackhardt (least) upper bound score.

network_connectedness(.data)

network_efficiency(.data)

network_upperbound(.data)

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

References

Krackhardt, David. 1994. Graph theoretical dimensions of informal organizations. In Carley and Prietula (eds) Computational Organizational Theory, Hillsdale, NJ: Lawrence Erlbaum Associates. Pp. 89-111.

Everett, Martin, and David Krackhardt. 2012. “A second look at Krackhardt's graph theoretical dimensions of informal organizations.” Social Networks, 34: 159-163. doi:10.1016/j.socnet.2011.10.006

See also

Examples

network_connectedness(ison_networkers)
#> [1] 1
1 - network_reciprocity(ison_networkers)
#> [1] 0.209
network_efficiency(ison_networkers)
#> [1] 0.0705
network_upperbound(ison_networkers)
#> [1] 1