Structural holes for one- and two-mode networks

These function provide different measures of the degree to which nodes fill structural holes, as outlined in Burt (1992).

node_bridges(object)

node_redundancy(object)

node_effsize(object)

node_efficiency(object)

node_constraint(object)

node_hierarchy(object)

Arguments

object

An object of a migraph-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

Functions

node_bridges: Returns nodes' bridge counts

node_redundancy: Returns nodes' redundancy

node_effsize: Returns nodes' effective size

node_efficiency: Returns nodes' efficiency

node_constraint: Returns nodes' constraint scores for one-mode networks according to Burt (1992) and for two-mode networks according to Hollway et al (2020).

node_hierarchy: Returns nodes' exposure to hierarchy, where only one or two contacts are the source of closure

References

#' Burt, Ronald S. 1992. Structural Holes: The Social Structure of Competition. Cambridge, MA: Harvard University Press.

Borgatti, Steven. 1997. “Structural Holes: Unpacking Burt’s Redundancy Measures” Connections 20(1):35-38.

Hollway, James, Jean-Frédéric Morin, and Joost Pauwelyn. 2020. "Structural conditions for novelty: the introduction of new environmental clauses to the trade regime complex." International Environmental Agreements: Politics, Law and Economics 20 (1): 61–83. doi:10.1007/s10784-019-09464-5 .

See also

Examples

node_bridges(ison_adolescents)
#>      Betty        Sue      Alice       Jane       Dale        Pam            
#> 1          1          0          0          0          1          + 2 others 
node_bridges(ison_southern_women)
#>      EVELYN       LAURA     THERESA      BRENDA   CHARLOTTE     FRANCES 
#> 0           0           0           0           0           0           
#>                      E9         E10         E11         E12         E13 
#> + 20 others 0           0           0           0           0           
#>         E14 
#> 0           
node_redundancy(ison_adolescents)
#>      Betty        Sue      Alice       Jane       Dale        Pam            
#> 6.00       2.50       2.50       4.00       3.33       4.00       + 2 others 
node_redundancy(ison_southern_women)
#>      EVELYN       LAURA     THERESA      BRENDA   CHARLOTTE     FRANCES 
#> 12.5        12.9        14.2        13.1        12.0        16.0        
#>                      E9         E10         E11         E12         E13 
#> + 20 others  8.00       10.00        7.50        9.43       12.00       
#>         E14 
#> 12.00       
node_effsize(ison_adolescents)
#>      Betty        Sue      Alice       Jane       Dale        Pam            
#> -5.000      1.500      1.500     -2.000     -0.333     -1.000     + 2 others 
node_effsize(ison_southern_women)
#>      EVELYN       LAURA     THERESA      BRENDA   CHARLOTTE     FRANCES 
#>  -4.50       -5.86       -6.25       -6.14       -8.00      -12.00      
#>                      E9         E10         E11         E12         E13 
#> + 20 others  4.00       -4.00       -3.50       -2.43       -8.00       
#>         E14 
#> -8.00       
node_efficiency(ison_adolescents)
#> [1] -5.000      0.375      0.375     -1.000     -0.111     -0.333     + 2 others
node_efficiency(ison_southern_women)
#>  [1] -0.562      -0.837      -0.781      -0.878      -2.000      -3.000     
#>  [7] + 20 others  0.333      -0.667      -0.875      -0.347      -2.000     
#> [13] -2.000     
node_constraint(ison_southern_women)
#>      EVELYN       LAURA     THERESA      BRENDA   CHARLOTTE     FRANCES 
#> 0.276       0.306       0.294       0.317       0.384       0.412       
#>                      E9         E10         E11         E12         E13 
#> + 20 others 0.161       0.327       0.308       0.292       0.432       
#>         E14 
#> 0.432       
node_hierarchy(ison_adolescents)
#>      Betty        Sue      Alice       Jane       Dale        Pam            
#> 0.00000    0.03059    0.02738    0.00587    0.03334    0.00612    + 2 others 
node_hierarchy(ison_southern_women)
#>      EVELYN       LAURA     THERESA      BRENDA   CHARLOTTE     FRANCES 
#> 0.0341      0.0359      0.0334      0.0301      0.0234      0.0207      
#>                      E9         E10         E11         E12         E13 
#> + 20 others 0.01497     0.01109     0.05673     0.00945     0.00490     
#>         E14 
#> 0.00490