Multiplex graph object of friends, social, and task ties (McFarland 2001)

Multiplex graph object of friends, social, and task ties (McFarland 2001)

data(ison_algebra)

Format

Multiplex tbl_graph object of friends, social, and task ties between 16 anonymous students.

Source

See also data(studentnets.M182, package = "NetData") Larger comprehensive data set publicly available, contact Daniel A. McFarland for details.

Details

Multiplex graph object of friends, social, and task ties between 16 #' anonymous students. M182 was an honors algebra class where researchers collected friendship, social, and task ties between 16 students. The edge attribute friend_ties contains friendship ties, where 2 = best friends, 1 = friend, and 0 is not a friend. social_ties consists of social interactions per hour, and task_ties consists of task interactions per hour.

References

McFarland, Daniel A. (2001) ``Student Resistance.'' American Journal of Sociology, 107(3), p 612-678.