One-mode EIES dataset (Freeman and Freeman 1979)

A directed, simple, named, weighted graph with 32 nodes and 440 edges. Nodes are academics and edges illustrate the communication patterns on an Electronic Information Exchange System among them. Node attributes include the number of citations (Citations) and the discipline of the researchers (Discipline). Edge weights illustrate the number of emails sent from one academic to another over the studied time period.

data(ison_networkers)

Format

tbl_graph network object. The network is directed, simple, named , and weighted. It contains 32 nodes and 440 edges as well as two node level attributes: Citations; Discipline.

Source

networkdata package

References

Freeman, S. C. and L. C. Freeman (1979). The networkers network: A study of the impact of a new communications medium on sociometric structure. Social Science Research Reports No 46. Irvine CA, University of California.

Wasserman S. and K. Faust (1994). Social Network Analysis: Methods and Applications. Cambridge University Press, Cambridge.