Marking nodes based on their properties

These functions return logical vectors the length of the nodes in a network identifying which hold certain properties or positions in the network.

node_is_cutpoint() marks nodes that cut or act as articulation points in a network, increasing the number of connected components when removed.

node_is_isolate() marks nodes that are isolates, with neither incoming nor outgoing ties.

node_is_core() marks nodes that are members of the network's core.

node_is_fold() marks nodes that are in a structural fold between two or more triangles that are only connected by that node.

node_is_mentor() marks a proportion of high indegree nodes as 'mentors' (see details)

node_is_infected() and node_is_exposed() marks nodes that are infected by a particular time point or exposed to a given (other) mark

node_is_random() marks one or more nodes at random.

node_is_max() and node_is_min() are more generally useful for converting the results from some node measure into a mark-class object. They can be particularly useful for highlighting which node or nodes are key because they minimise or, more often, maximise some measure.

node_is_isolate(.data)

node_is_cutpoint(.data)

node_is_core(.data)

node_is_fold(.data)

node_is_mentor(.data, elites = 0.1)

node_is_latent(diff_model, time = 0)

node_is_infected(diff_model, time = 0)

node_is_recovered(diff_model, time = 0)

node_is_exposed(.data, mark)

node_is_random(.data, size = 1)

node_is_max(node_measure, ranks = 1)

node_is_min(node_measure, ranks = 1)

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

elites

The proportion of nodes to be selected as mentors. By default this is set at 0.1. This means that the top 10% of nodes in terms of degree, or those equal to the highest rank degree in the network, whichever is the higher, will be used to select the mentors.

Note that if nodes are equidistant from two mentors, they will choose one at random. If a node is without a path to a mentor, for example because they are an isolate, a tie to themselves (a loop) will be created instead. Note that this is a different default behaviour than that described in Valente and Davis (1999).

diff_model

A valid network diffusion model, as created by as_diffusion() or play_diffusion().

time

A time point until which infections/adoptions should be identified. By default time = 0.

mark

vector denoting which nodes are infected

size

The number of nodes to select (as TRUE).

node_measure

An object created by a node_ measure.

ranks

The number of ranks of max or min to return. For example, ranks = 3 will return TRUE for nodes with scores equal to any of the top (or, for node_is_min(), bottom) three scores. By default, ranks = 1.

Exposed

node_is_exposed() is similar to node_exposure(), but returns a mark (TRUE/FALSE) vector indicating which nodes are currently exposed to the diffusion content. This diffusion content can be expressed in the 'mark' argument. If no 'mark' argument is provided, and '.data' is a diff_model object, then the function will return nodes exposure to the seed nodes in that diffusion.

References

Valente, Thomas, and Rebecca Davis. 1999. "Accelerating the Diffusion of Innovations Using Opinion Leaders", Annals of the American Academy of Political and Social Science 566: 56-67.

See also

Examples

node_is_isolate(ison_brandes)
#>   V1    V2    V3    V4    V5    V6    V7    V8    V9    V10   V11  
#> 1 FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
node_is_cutpoint(ison_brandes)
#>   V1    V2    V3    V4    V5    V6    V7    V8    V9    V10   V11  
#> 1 FALSE FALSE TRUE  TRUE  FALSE FALSE FALSE FALSE TRUE  FALSE FALSE
node_is_core(ison_brandes)
#>   V1    V2    V3    V4    V5    V6    V7    V8    V9    V10   V11  
#> 1 FALSE FALSE TRUE  TRUE  FALSE FALSE FALSE FALSE TRUE  FALSE FALSE
node_is_fold(create_explicit(A-B, B-C, A-C, C-D, C-E, D-E))
#>   A     B     C     D     E    
#> 1 FALSE FALSE TRUE  FALSE FALSE
  # To mark nodes that are latent by a particular time point
  node_is_latent(play_diffusion(create_tree(6), latency = 1), time = 1)
#>   V1    V2    V3    V4    V5    V6   
#> 1 FALSE TRUE  TRUE  FALSE FALSE FALSE
  # To mark nodes that are infected by a particular time point
  node_is_infected(play_diffusion(create_tree(6)), time = 1)
#>   V1    V2    V3    V4    V5    V6   
#> 1 TRUE  TRUE  TRUE  FALSE FALSE FALSE
  # To mark nodes that are recovered by a particular time point
  node_is_recovered(play_diffusion(create_tree(6), recovery = 0.5), time = 3)
#>   V1    V2    V3    V4    V5    V6   
#> 1 TRUE  FALSE FALSE FALSE FALSE FALSE
  # To mark which nodes are currently exposed
  (expos <- node_is_exposed(manynet::create_tree(14), mark = c(1,3)))
#>   V1    V2    V3    V4    V5    V6    V7    V8    V9    V10   V11   V12   V13  
#> 1 FALSE TRUE  FALSE FALSE FALSE TRUE  TRUE  FALSE FALSE FALSE FALSE FALSE FALSE
#> # ... with 1 more values from this nodeset unprinted. Use `print(..., n = Inf)` to print all values.
  which(expos)
#> [1] 2 6 7
node_is_random(ison_brandes, 2)
#>   V1    V2    V3    V4    V5    V6    V7    V8    V9    V10   V11  
#> 1 TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE
node_is_max(node_degree(ison_brandes))
#>   V1    V2    V3    V4    V5    V6    V7    V8    V9    V10   V11  
#> 1 FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE FALSE
node_is_min(node_degree(ison_brandes))
#>   V1    V2    V3    V4    V5    V6    V7    V8    V9    V10   V11  
#> 1 TRUE  TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE  TRUE