One-mode interstate trade relations and two-mode state membership in IGOs (COW)

One-mode interstate trade relations and two-mode state membership in IGOs (COW)

data(mpn_cow_trade)

data(mpn_cow_igo)

Format

## # A tbl_graph: 116 nodes and 11489 edges
## #
## # A directed simple graph with 1 component
## #
## # Node Data: 116 × 1 (active)
##   name                    
##   <chr>                   
## 1 United States of America
## 2 Canada                  
## 3 Cuba                    
## 4 Dominican Republic      
## 5 Jamaica                 
## 6 Trinidad and Tobago     
## # … with 110 more rows
## #
## # Edge Data: 11,489 × 3
##    from    to  weight
##   <int> <int>   <dbl>
## 1     1     2 180387 
## 2     1     3    587.
## 3     1     4   5511.
## # … with 11,486 more rows

## # A tbl_graph: 156 nodes and 862 edges
## #
## # A bipartite simple graph with 1 component
## #
## # Node Data: 156 × 2 (active)
##   type  name       
##   <lgl> <chr>      
## 1 FALSE Afghanistan
## 2 FALSE Albania    
## 3 FALSE Algeria    
## 4 FALSE Angola     
## 5 FALSE Argentina  
## 6 FALSE Australia  
## # … with 150 more rows
## #
## # Edge Data: 862 × 2
##    from    to
##   <int> <int>
## 1     1   121
## 2     1   132
## 3     1   139
## # … with 859 more rows

Source

The Correlates of War Project. 2012. Trade.

Barbieri, Katherine and Omar Keshk. 2012. Correlates of War Project Trade Data Set Codebook, Version 3.0..

The Correlates of War Project. 2019. Intergovernmental Organization v3.

Details

mpn_cow_trade is a one-mode matrix representing the trade relations between 116 states. The data is derived from the Correlates of War Project (COW) Trade Dataset (v3.0), which contains the annual dyadic and national trade figures for states (listed in COW) between 1870 to 2009. This network is based only on the dyadic trade figures in 2009 for the 116 states listed in Appendix 7.1 in Multimodal Political Networks (Knoke et al., 2021). The value in each cell of the matrix represents the value of exports from the 116 row states to the 116 column states.

Two-mode matrix representing the membership of 116 states to 40 intergovernmental organizations (IGOs). The data is derived from the Correlates of War Project (COW) Intergovernmental Organizations Dataset (v3.0), which contains information about intergovernmental organizations from 1815-2014, such as founding year and membership. This network contains only a subset of the states and IGOs listed in COW, with 116 states listed in Appendix 7.1 in Multimodal Political Networks and 40 IGOs from Table 7.1 in Multimodal Political Networks that also overlap with the COW dataset (Knoke et al., 2021).

References

Barbieri, Katherine, Omar M. G. Keshk, and Brian Pollins. 2009. “TRADING DATA: Evaluating our Assumptions and Coding Rules.” Conflict Management and Peace Science 26(5): 471-491. doi:10.1177/0738894209343887 .

Knoke, David, Mario Diani, James Hollway, and Dimitris C Christopoulos. 2021. Multimodal Political Networks. Cambridge University Press. Cambridge University Press.

Pevehouse, Jon C.W., Timothy Nordstron, Roseanne W McManus, Anne Spencer Jamison. 2020. “Tracking Organizations in the World: The Correlates of War IGO Version 3.0 datasets”. Journal of Peace Research 57(3): 492-503. doi:10.1177/0022343319881175 .