Two-mode European Values Survey, 1990 and 2008 (EVS 2020)

6 two-mode matrices containing individuals' memberships to 14 different types of associations in three countries (Italy, Germany, and the UK) in 1990 and 2008. The Italy data has 658 respondents in 1990 and 540 in 2008. The Germany data has 1369 respondents in 1990 and 503 in 2008. The UK data has 738 respondents in 1990 and 664 in 2008.

data(mpn_IT_1990)

data(mpn_IT_1990)

data(mpn_IT_2008)

data(mpn_DE_1990)

data(mpn_DE_2008)

data(mpn_UK_1990)

data(mpn_UK_2008)

Format

tbl_graph object based on an association matrix with 14 columns:

Welfare

1 if individual associated

Religious

1 if individual associated

Education.culture

1 if individual associated

Unions

1 if individual associated

Parties

1 if individual associated

Local.political.groups

1 if individual associated

Human.rights

1 if individual associated

Environmental.animal

1 if individual associated

Professional

1 if individual associated

Youth

1 if individual associated

Sports

1 if individual associated

Women

1 if individual associated

Peace

1 if individual associated

Health

1 if individual associated

An object of class tbl_graph (inherits from igraph) of length 672.

An object of class tbl_graph (inherits from igraph) of length 554.

An object of class tbl_graph (inherits from igraph) of length 1383.

An object of class tbl_graph (inherits from igraph) of length 517.

An object of class tbl_graph (inherits from igraph) of length 752.

An object of class tbl_graph (inherits from igraph) of length 678.

Source

Knoke, David, Mario Diani, James Hollway, and Dimitris C Christopoulos. 2021. Multimodal Political Networks. Cambridge University Press. Cambridge University Press.

References

EVS (2020). European Values Study Longitudinal Data File 1981-2008 (EVS 1981-2008). GESIS Data Archive, Cologne. ZA4804 Data file Version 3.1.0, doi:10.4232/1.13486 .