One-mode EU policy influence network, June 2004 (Christopoulos 2006)

Network of anonymised actors reacting to the Ryanair/Charleroi decision of the EU Commission in February 2004. The relationships mapped comprise an account of public records of interaction supplemented with the cognitive network of key informants. Examination of relevant communiques, public statements and a number of off-the-record interviews provides confidence that the network mapped closely approximated interactions between 29 January and 12 February 2004. The time point mapped is at the height of influence and interest intermediation played by actors in the AER, a comparatively obscure body representing the interests of a number of European regional bodies at the EU institutions.

data(mpn_ryanair)

Format

#> # A labelled, weighted, directed tbl_graph with 20 nodes and 177 arcs
#> # A tibble: 20 x 1
#>   name          
#>   <chr>         
#> 1 1 AER         
#> 2 2 AER         
#> 3 5 AER/COR     
#> 4 7 RYANAIR     
#> 5 8 DG TRANSPORT
#> 6 9 COR         
#> # i 14 more rows
#> # A tibble: 177 x 3
#>    from    to weight
#>   <int> <int>  <dbl>
#> 1     1     2      1
#> 2     1     3      1
#> 3     1     4      1
#> 4     1     5      1
#> 5     1     6      1
#> 6     1     7      1
#> # i 171 more rows

Source

Christopoulos, Dimitrios C. 2006. “Relational Attributes of Political Entrepreneurs: a Network Perspective.” Journal of European Public Policy 13(5): 757–78. doi:10.1080/13501760600808964 .

Knoke, David, Mario Diani, James Hollway, and Dimitris C Christopoulos. 2021. Multimodal Political Networks. Cambridge University Press. Cambridge University Press.