Censuses of motifs at the network level

These functions include ways to take a census of the positions of nodes in a network:

network_dyad_census() returns a census of dyad motifs in a network.

network_triad_census() returns a census of triad motifs in a network.

network_mixed_census() returns a census of triad motifs that span a one-mode and a two-mode network.

network_dyad_census(.data)

network_triad_census(.data)

network_mixed_census(.data, object2)

Source

Alejandro Espinosa 'netmem'

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

object2

A second, two-mode migraph-consistent object.

References

Davis, James A., and Samuel Leinhardt. 1967. “The Structure of Positive Interpersonal Relations in Small Groups.” 55.

Hollway, James, Alessandro Lomi, Francesca Pallotti, and Christoph Stadtfeld. 2017. “Multilevel Social Spaces: The Network Dynamics of Organizational Fields.” Network Science 5(2): 187–212. doi:10.1017/nws.2017.8

See also

Examples

network_dyad_census(manynet::ison_algebra)
#> # A tibble: 1 × 3
#>   Mutual Asymmetric  Null
#>    <dbl>      <dbl> <dbl>
#> 1     68          8    44
network_triad_census(manynet::ison_adolescents)
#> # A tibble: 1 × 6
#>   `003` `012` `102` `201` `210` `300`
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1    13     0    29    11     0     3
marvel_friends <- manynet::to_unsigned(manynet::ison_marvel_relationships, "positive")
(mixed_cen <- network_mixed_census(marvel_friends, manynet::ison_marvel_teams))
#> # A tibble: 1 × 10
#>    `22`  `21`  `20`  `12` `11D` `11U`  `10`  `02`  `01`   `00`
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1  1137     0  1702  7064     0     0 22774 30856     0 130765