Constraint for one- and two-mode networks

This function measures constraint for both one-mode and two-mode networks. For one-mode networks, the function wraps the implementation of Ron Burt's measure in {igraph}. For two-mode networks, the function employs the extension outlined in Hollway et al. (2020).

node_constraint(object, nodes = V(object), weights = NULL)

Arguments

object

A matrix, igraph graph, or tidygraph object.

nodes

The vertices for which the constraint will be calculated. Defaults to all vertices.

weights

The weights of the edges. If this is NULL and there is a weight edge attribute this is used. If there is no such edge attribute all edges will have the same weight.

Value

A named vector (one-mode) or a list of two named vectors ($nodes1, $nodes2).

References

Hollway, James, Jean-Frédéric Morin, and Joost Pauwelyn. 2020. "Structural conditions for novelty: the introduction of new environmental clauses to the trade regime complex." International Environmental Agreements: Politics, Law and Economics 20 (1): 61–83.

See also

Examples

node_constraint(ison_southern_women)
#>      EVELYN       LAURA     THERESA      BRENDA   CHARLOTTE     FRANCES 
#> 0.276       0.306       0.294       0.317       0.384       0.412       
#>                      E9         E10         E11         E12         E13 
#> + 20 others 0.161       0.327       0.308       0.292       0.432       
#>         E14 
#> 0.432