Helper functions for measuring over splits of networks

Helper functions for measuring over splits of networks

over_waves(
  .data,
  FUN,
  ...,
  attribute = "wave",
  strategy = "sequential",
  verbose = FALSE
)

over_time(
  .data,
  FUN,
  ...,
  attribute = "time",
  slice = NULL,
  strategy = "sequential",
  verbose = FALSE
)

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

FUN

A function to run over all splits.

...

Further arguments to be passed on to FUN.

attribute

A string naming the attribute to be split upon.

strategy

If {furrr} is installed, then multiple cores can be used to accelerate the function. By default "sequential", but if multiple cores available, then "multisession" or "multicore" may be useful. Generally this is useful only when times > 1000. See {furrr} for more.

verbose

Whether the function should report on its progress. By default FALSE. See {progressr} for more.

slice

Optionally, a vector of specific slices. Otherwise all observed slices will be returned.

Functions

over_waves(): Runs a function, e.g. a measure, over waves of a panel network

over_time(): Runs a function, e.g. a measure, over time slices of a dynamic network