Census by nodes or clusters

These functions include ways to take a census of the positions of nodes in a network. These include a triad census based on the triad profile of nodes, but also a tie census based on the particular tie partners of nodes. Included also are group census functions for summarising the profiles of clusters of nodes in a network.

node_tie_census(object)

node_triad_census(object)

node_quad_census(object)

group_tie_census(object, clusters, decimals = 2)

group_triad_census(object, clusters, decimals = 2)

Arguments

object
A migraph-consistent object.

clusters
a vector of cluster assignment.

decimals
Number of decimal points to round to.

Examples

task_eg <- to_named(to_uniplex(ison_m182, "task_tie"))
(tie_cen <- node_tie_census(task_eg))
#>           fromDerrick fromKiara fromTamika fromKent fromRachel fromBraden
#> Derrick          0.00      0.00       0.00     0.00       0.30       0.00
#> Kiara            0.00      0.00       0.00     0.00       0.00       0.00
#> Tamika           0.00      0.00       0.00     0.00       0.15       0.75
#> Kent             0.00      0.00       0.00     0.00       0.00       0.00
#> Rachel           0.30      0.00       0.15     0.00       0.00       0.45
#> Braden           0.00      0.00       0.60     0.00       0.75       0.00
#> Isla             0.00      0.75       0.00     0.00       0.00       0.00
#> Chris            0.00      0.30       0.00     0.00       0.00       0.00
#> Bettie           0.30      0.00       0.00     0.00       0.00       0.00
#> Celia            0.30      0.00       0.00     0.15       0.00       0.15
#> Heather          0.00      0.00       0.30     0.00       2.10       0.30
#> Mae              0.15      0.00       0.00     0.00       0.00       0.00
#> Nicolas          0.00      0.90       0.00     0.00       0.15       0.00
#> Micah            0.00      0.15       0.00     0.00       0.00       0.00
#> Johnathan        0.15      0.15       0.00     0.00       0.00       0.00
#> Adriana          5.10      4.80       0.30     0.90       1.80       0.90
#>           fromIsla fromChris fromBettie fromCelia fromHeather fromMae
#> Derrick       0.00      0.00       0.15      0.15        0.00    0.00
#> Kiara         1.20      0.45       0.00      0.00        0.00    0.00
#> Tamika        0.00      0.00       0.00      0.00        0.15    0.00
#> Kent          0.00      0.00       0.00      0.15        0.00    0.00
#> Rachel        0.00      0.00       0.00      0.00        1.80    0.00
#> Braden        0.00      0.00       0.00      0.30        0.45    0.00
#> Isla          0.00      0.30       0.00      0.00        0.00    0.00
#> Chris         0.15      0.00       0.00      0.00        0.00    0.00
#> Bettie        0.00      0.00       0.00      0.45        0.00    2.40
#> Celia         0.00      0.00       0.30      0.00        0.00    0.45
#> Heather       0.00      0.00       0.00      0.00        0.00    0.00
#> Mae           0.00      0.00       2.70      0.75        0.00    0.00
#> Nicolas       0.30      0.75       0.00      0.00        0.00    0.00
#> Micah         0.00      0.15       0.00      0.00        0.00    0.00
#> Johnathan     0.00      0.00       0.15      0.30        0.00    0.75
#> Adriana       2.85      2.40       2.40      5.40        1.05    4.50
#>           fromNicolas fromMicah fromJohnathan fromAdriana toDerrick toKiara
#> Derrick          0.00      0.00          0.15        4.65      0.00    0.00
#> Kiara            1.05      0.30          0.00        4.80      0.00    0.00
#> Tamika           0.00      0.00          0.00        0.75      0.00    0.00
#> Kent             0.00      0.00          0.00        1.05      0.00    0.00
#> Rachel           0.00      0.15          0.00        2.40      0.30    0.00
#> Braden           0.00      0.15          0.00        1.20      0.00    0.00
#> Isla             0.30      0.15          0.00        2.40      0.00    1.20
#> Chris            0.60      0.30          0.00        1.95      0.00    0.45
#> Bettie           0.00      0.00          0.15        2.85      0.15    0.00
#> Celia            0.00      0.00          0.30        5.40      0.15    0.00
#> Heather          0.00      0.00          0.00        1.20      0.00    0.00
#> Mae              0.00      0.00          0.45        5.25      0.00    0.00
#> Nicolas          0.00      0.00          0.00        3.90      0.00    1.05
#> Micah            0.00      0.00          0.00        3.90      0.00    0.30
#> Johnathan        0.00      0.00          0.00        5.55      0.15    0.00
#> Adriana          4.50      3.15          6.60        0.00      4.65    4.80
#>           toTamika toKent toRachel toBraden toIsla toChris toBettie toCelia
#> Derrick       0.00   0.00     0.30     0.00   0.00    0.00     0.30    0.30
#> Kiara         0.00   0.00     0.00     0.00   0.75    0.30     0.00    0.00
#> Tamika        0.00   0.00     0.15     0.60   0.00    0.00     0.00    0.00
#> Kent          0.00   0.00     0.00     0.00   0.00    0.00     0.00    0.15
#> Rachel        0.15   0.00     0.00     0.75   0.00    0.00     0.00    0.00
#> Braden        0.75   0.00     0.45     0.00   0.00    0.00     0.00    0.15
#> Isla          0.00   0.00     0.00     0.00   0.00    0.15     0.00    0.00
#> Chris         0.00   0.00     0.00     0.00   0.30    0.00     0.00    0.00
#> Bettie        0.00   0.00     0.00     0.00   0.00    0.00     0.00    0.30
#> Celia         0.00   0.15     0.00     0.30   0.00    0.00     0.45    0.00
#> Heather       0.15   0.00     1.80     0.45   0.00    0.00     0.00    0.00
#> Mae           0.00   0.00     0.00     0.00   0.00    0.00     2.40    0.45
#> Nicolas       0.00   0.00     0.00     0.00   0.30    0.60     0.00    0.00
#> Micah         0.00   0.00     0.15     0.15   0.15    0.30     0.00    0.00
#> Johnathan     0.00   0.00     0.00     0.00   0.00    0.00     0.15    0.30
#> Adriana       0.75   1.05     2.40     1.20   2.40    1.95     2.85    5.40
#>           toHeather toMae toNicolas toMicah toJohnathan toAdriana
#> Derrick         0.0  0.15      0.00    0.00        0.15      5.10
#> Kiara           0.0  0.00      0.90    0.15        0.15      4.80
#> Tamika          0.3  0.00      0.00    0.00        0.00      0.30
#> Kent            0.0  0.00      0.00    0.00        0.00      0.90
#> Rachel          2.1  0.00      0.15    0.00        0.00      1.80
#> Braden          0.3  0.00      0.00    0.00        0.00      0.90
#> Isla            0.0  0.00      0.30    0.00        0.00      2.85
#> Chris           0.0  0.00      0.75    0.15        0.00      2.40
#> Bettie          0.0  2.70      0.00    0.00        0.15      2.40
#> Celia           0.0  0.75      0.00    0.00        0.30      5.40
#> Heather         0.0  0.00      0.00    0.00        0.00      1.05
#> Mae             0.0  0.00      0.00    0.00        0.75      4.50
#> Nicolas         0.0  0.00      0.00    0.00        0.00      4.50
#> Micah           0.0  0.00      0.00    0.00        0.00      3.15
#> Johnathan       0.0  0.45      0.00    0.00        0.00      6.60
#> Adriana         1.2  5.25      3.90    3.90        5.55      0.00
(triad_cen <- node_triad_census(task_eg))
#>           003 012 102 021D 021U 021C 111D 111U 030T 030C 201 120D 120U 120C 210
#> Derrick    36   9  45    0    0    0    0    1    0    0   3    0    0    0   4
#> Kiara      45  10  38    0    0    0    1    1    0    0   1    0    0    0   3
#> Tamika     55   0  44    0    0    0    0    0    0    0   0    0    0    0   0
#> Kent      102   0   3    0    0    0    0    0    0    0   0    0    0    0   0
#> Rachel     55   0  37    0    0    0    2    1    0    0   4    0    0    0   0
#> Braden     65  19  10    0    0    0    5    0    0    0   3    1    0    0   2
#> Isla       91   1  13    0    0    0    0    0    0    0   0    0    0    0   0
#> Chris      91   1  13    0    0    0    0    0    0    0   0    0    0    0   0
#> Bettie     90   1  14    0    0    0    0    0    0    0   0    0    0    0   0
#> Celia      86   1  18    0    0    0    0    0    0    0   0    0    0    0   0
#> Heather    95   0  10    0    0    0    0    0    0    0   0    0    0    0   0
#> Mae        90   1  14    0    0    0    0    0    0    0   0    0    0    0   0
#> Nicolas    93   1  11    0    0    0    0    0    0    0   0    0    0    0   0
#> Micah      90   3  12    0    0    0    0    0    0    0   0    0    0    0   0
#> Johnathan  88   2  15    0    0    0    0    0    0    0   0    0    0    0   0
#> Adriana     0   0   0    0    0    0    0    0    0    0  73    0    0    0   6
#>           300
#> Derrick     7
#> Kiara       6
#> Tamika      6
#> Kent        0
#> Rachel      6
#> Braden      0
#> Isla        0
#> Chris       0
#> Bettie      0
#> Celia       0
#> Heather     0
#> Mae         0
#> Nicolas     0
#> Micah       0
#> Johnathan   0
#> Adriana    26
(quad_cen <- node_quad_census(southern_women))
#>             E4  I41  I40  H4 L42 L41 L40 U42 U41 Y43 Y41  04
#> EVELYN    1402 1046  474 202 397 692 148 200  78  56 190  75
#> LAURA     1539  986  579 202 323 659 182 134  85  35 168  68
#> THERESA   1499  975  434 147 364 784  91 217  77  56 225  91
#> BRENDA    1559  968  562 198 320 682 179 134  82  35 170  71
#> CHARLOTTE 1945  763  972 171 117 419 359  30  89   4  70  21
#> FRANCES   2063  652  944 126 102 502 269  36 102   4 130  30
#> ELEANOR   2122  598  930  98  93 546 224  42 114   4 156  33
#> PEARL     2313  439 1089  65  42 448 252  28 112   1 154  17
#> RUTH      2180  549  918  73  80 577 178  56 123   4 190  32
#> VERNE     2166  561  920  82  83 567 190  53 118   4 184  32
#> MYRA      2111  609  926 116  94 529 239  43 100   4 158  31
#> KATHERINE 1707  897  682 192 239 613 211  97  84  20 164  54
#> SYLVIA    1614  930  544 170 299 718 142 152  82  35 200  74
#> NORA      1387 1058  438 215 403 700 199 209  89  56 146  60
#> HELEN     1503 1021  582 219 327 623 215 144  89  35 148  54
#> DOROTHY   2111  609  926 116  94 529 239  43 100   4 158  31
#> OLIVIA    2341  425 1282  97  18 233 437  10  57   0  58   2
#> FLORA     2341  425 1282  97  18 233 437  10  57   0  58   2
#> E1        2113  629 1165 108  64 325 376   8 105   1  51  15
#> E2        2123  619 1172  96  63 332 359   8 114   1  57  16
#> E3        1554 1014  772 214 284 528 274  52 116  20  78  54
#> E4        1899  804 1038 147 123 384 339  21 117   4  60  24
#> E5        1216 1154  582 264 487 594 226 104 112  56  84  81
#> E6        1239 1159  602 206 456 597 183 145 146  56 100  71
#> E7         994 1200  432 174 630 666 114 270 156 120 114  90
#> E8         603 1063  170 139 975 724  43 507  85 364 131 156
#> E9         671 1210  344 259 884 573 125 344 137 220 101  92
#> E10       1520 1043  792 224 291 498 288  48 118  20  67  51
#> E11       1767  936 1094 207 131 244 415  29  89   4  36   8
#> E12       1343 1126  695 252 391 533 262  70 119  35  70  64
#> E13       1924  777 1014 148 124 414 338  16 112   4  61  28
#> E14       1924  777 1014 148 124 414 338  16 112   4  61  28
group_tie_census(task_eg, cutree(cluster_structural_equivalence(task_eg), 4))
#>         fromDerrick fromKiara fromTamika fromKent fromRachel fromBraden
#> Block 1        0.08       0.2       0.00     0.01       0.04       0.01
#> Block 2        0.15       0.0       0.07     0.00       0.07       0.60
#> Block 3        0.00       0.0       0.45     0.00       1.43       0.15
#> Block 4        5.10       4.8       0.30     0.90       1.80       0.90
#>         fromIsla fromChris fromBettie fromCelia fromHeather fromMae fromNicolas
#> Block 1     0.15      0.15        0.3      0.16        0.00    0.33        0.18
#> Block 2     0.00      0.00        0.0      0.00        0.98    0.00        0.00
#> Block 3     0.00      0.00        0.0      0.15        0.22    0.00        0.00
#> Block 4     2.85      2.40        2.4      5.40        1.05    4.50        4.50
#>         fromMicah fromJohnathan fromAdriana toDerrick toKiara toTamika toKent
#> Block 1      0.07           0.1        3.79      0.04    0.27     0.00   0.01
#> Block 2      0.07           0.0        1.57      0.15    0.00     0.07   0.00
#> Block 3      0.07           0.0        1.20      0.00    0.00     0.45   0.00
#> Block 4      3.15           6.6        0.00      4.65    4.80     0.75   1.05
#>         toRachel toBraden toIsla toChris toBettie toCelia toHeather toMae
#> Block 1     0.04     0.04   0.14    0.12     0.30    0.14      0.00  0.37
#> Block 2     0.07     0.68   0.00    0.00     0.00    0.00      1.20  0.00
#> Block 3     1.12     0.22   0.00    0.00     0.00    0.07      0.15  0.00
#> Block 4     2.40     1.20   2.40    1.95     2.85    5.40      1.20  5.25
#>         toNicolas toMicah toJohnathan toAdriana
#> Block 1      0.18    0.03        0.14      3.87
#> Block 2      0.07    0.00        0.00      1.05
#> Block 3      0.00    0.00        0.00      0.98
#> Block 4      3.90    3.90        5.55      0.00
group_triad_census(task_eg, cutree(cluster_regular_equivalence(task_eg), 4))
#>           003  012   102 021D 021U 021C 111D 111U 030T 030C   201 120D 120U
#> Block 1 36.00 9.00 45.00    0    0    0 0.00 1.00    0    0  3.00 0.00    0
#> Block 2 51.67 3.33 39.67    0    0    0 1.00 0.67    0    0  1.67 0.00    0
#> Block 3 89.18 2.73 12.09    0    0    0 0.45 0.00    0    0  0.27 0.09    0
#> Block 4  0.00 0.00  0.00    0    0    0 0.00 0.00    0    0 73.00 0.00    0
#>         120C  210 300
#> Block 1    0 4.00   7
#> Block 2    0 1.00   6
#> Block 3    0 0.18   0
#> Block 4    0 6.00  26