Census by nodes or clusters

These functions include ways to take a census of the positions of nodes in a network. These include a triad census based on the triad profile of nodes, but also a tie census based on the particular tie partners of nodes. Included also are group census functions for summarising the profiles of clusters of nodes in a network.

node_tie_census(object)

node_triad_census(object)

node_quad_census(object)

group_tie_census(object, clusters, decimals = 2)

group_triad_census(object, clusters, decimals = 2)

Arguments

object

An object of a migraph-consistent class:

matrix, from base R

edgelist, a data frame from base R or tibble from tibble

igraph, from the igraph package

network, from the network package

tbl_graph, from the tidygraph package

clusters

a vector of cluster assignment.

decimals

Number of decimal points to round to.

Details

The quad census uses the {oaqc} package to do the heavy lifting of counting the number of each orbits. See vignette('oaqc'). However, our function relabels some of the motifs to avoid conflicts and improve some consistency with other census-labelling practices. The letter-number pairing of these labels indicate the number and configuration of ties. For now, we offer a rough translation:

migraph	Ortmann and Brandes

E4	co-K4

I40, I41	co-diamond

H4	co-C4

L42, L41, L40	co-paw

D42, D40	co-claw

U42, U41	P4

Y43, Y41	claw

P43, P42, P41	paw

04	C4

Z42, Z43	diamond

X4	K4

Functions

node_quad_census: Returns a census of nodes' positions in motifs of four nodes.

References

Ortmann, Mark, and Ulrik Brandes. 2017. “Efficient Orbit-Aware Triad and Quad Census in Directed and Undirected Graphs.” Applied Network Science 2(1):13.

Examples

task_eg <- to_named(to_uniplex(ison_algebra, "task_tie"))
(tie_cen <- node_tie_census(task_eg))
#>            fromGoldie fromTroy fromJermaine fromJacob fromLeona fromVeronica
#> Goldie           0.00     0.00         0.00      0.00      0.30         0.00
#> Troy             0.00     0.00         0.00      0.00      0.00         0.00
#> Jermaine         0.00     0.00         0.00      0.00      0.15         0.75
#> Jacob            0.00     0.00         0.00      0.00      0.00         0.00
#> Leona            0.30     0.00         0.15      0.00      0.00         0.45
#> Veronica         0.00     0.00         0.60      0.00      0.75         0.00
#> Latoya           0.00     0.75         0.00      0.00      0.00         0.00
#> Dave             0.00     0.30         0.00      0.00      0.00         0.00
#> Jaxson           0.30     0.00         0.00      0.00      0.00         0.00
#> Marguerite       0.30     0.00         0.00      0.15      0.00         0.15
#> Janet            0.00     0.00         0.30      0.00      2.10         0.30
#> Arya             0.15     0.00         0.00      0.00      0.00         0.00
#> Priscilla        0.00     0.90         0.00      0.00      0.15         0.00
#> Essie            0.00     0.15         0.00      0.00      0.00         0.00
#> Grayson          0.15     0.15         0.00      0.00      0.00         0.00
#> Roberto          5.10     4.80         0.30      0.90      1.80         0.90
#>            fromLatoya fromDave fromJaxson fromMarguerite fromJanet fromArya
#> Goldie           0.00     0.00       0.15           0.15      0.00     0.00
#> Troy             1.20     0.45       0.00           0.00      0.00     0.00
#> Jermaine         0.00     0.00       0.00           0.00      0.15     0.00
#> Jacob            0.00     0.00       0.00           0.15      0.00     0.00
#> Leona            0.00     0.00       0.00           0.00      1.80     0.00
#> Veronica         0.00     0.00       0.00           0.30      0.45     0.00
#> Latoya           0.00     0.30       0.00           0.00      0.00     0.00
#> Dave             0.15     0.00       0.00           0.00      0.00     0.00
#> Jaxson           0.00     0.00       0.00           0.45      0.00     2.40
#> Marguerite       0.00     0.00       0.30           0.00      0.00     0.45
#> Janet            0.00     0.00       0.00           0.00      0.00     0.00
#> Arya             0.00     0.00       2.70           0.75      0.00     0.00
#> Priscilla        0.30     0.75       0.00           0.00      0.00     0.00
#> Essie            0.00     0.15       0.00           0.00      0.00     0.00
#> Grayson          0.00     0.00       0.15           0.30      0.00     0.75
#> Roberto          2.85     2.40       2.40           5.40      1.05     4.50
#>            fromPriscilla fromEssie fromGrayson fromRoberto toGoldie toTroy
#> Goldie              0.00      0.00        0.15        4.65     0.00   0.00
#> Troy                1.05      0.30        0.00        4.80     0.00   0.00
#> Jermaine            0.00      0.00        0.00        0.75     0.00   0.00
#> Jacob               0.00      0.00        0.00        1.05     0.00   0.00
#> Leona               0.00      0.15        0.00        2.40     0.30   0.00
#> Veronica            0.00      0.15        0.00        1.20     0.00   0.00
#> Latoya              0.30      0.15        0.00        2.40     0.00   1.20
#> Dave                0.60      0.30        0.00        1.95     0.00   0.45
#> Jaxson              0.00      0.00        0.15        2.85     0.15   0.00
#> Marguerite          0.00      0.00        0.30        5.40     0.15   0.00
#> Janet               0.00      0.00        0.00        1.20     0.00   0.00
#> Arya                0.00      0.00        0.45        5.25     0.00   0.00
#> Priscilla           0.00      0.00        0.00        3.90     0.00   1.05
#> Essie               0.00      0.00        0.00        3.90     0.00   0.30
#> Grayson             0.00      0.00        0.00        5.55     0.15   0.00
#> Roberto             4.50      3.15        6.60        0.00     4.65   4.80
#>            toJermaine toJacob toLeona toVeronica toLatoya toDave toJaxson
#> Goldie           0.00    0.00    0.30       0.00     0.00   0.00     0.30
#> Troy             0.00    0.00    0.00       0.00     0.75   0.30     0.00
#> Jermaine         0.00    0.00    0.15       0.60     0.00   0.00     0.00
#> Jacob            0.00    0.00    0.00       0.00     0.00   0.00     0.00
#> Leona            0.15    0.00    0.00       0.75     0.00   0.00     0.00
#> Veronica         0.75    0.00    0.45       0.00     0.00   0.00     0.00
#> Latoya           0.00    0.00    0.00       0.00     0.00   0.15     0.00
#> Dave             0.00    0.00    0.00       0.00     0.30   0.00     0.00
#> Jaxson           0.00    0.00    0.00       0.00     0.00   0.00     0.00
#> Marguerite       0.00    0.15    0.00       0.30     0.00   0.00     0.45
#> Janet            0.15    0.00    1.80       0.45     0.00   0.00     0.00
#> Arya             0.00    0.00    0.00       0.00     0.00   0.00     2.40
#> Priscilla        0.00    0.00    0.00       0.00     0.30   0.60     0.00
#> Essie            0.00    0.00    0.15       0.15     0.15   0.30     0.00
#> Grayson          0.00    0.00    0.00       0.00     0.00   0.00     0.15
#> Roberto          0.75    1.05    2.40       1.20     2.40   1.95     2.85
#>            toMarguerite toJanet toArya toPriscilla toEssie toGrayson toRoberto
#> Goldie             0.30     0.0   0.15        0.00    0.00      0.15      5.10
#> Troy               0.00     0.0   0.00        0.90    0.15      0.15      4.80
#> Jermaine           0.00     0.3   0.00        0.00    0.00      0.00      0.30
#> Jacob              0.15     0.0   0.00        0.00    0.00      0.00      0.90
#> Leona              0.00     2.1   0.00        0.15    0.00      0.00      1.80
#> Veronica           0.15     0.3   0.00        0.00    0.00      0.00      0.90
#> Latoya             0.00     0.0   0.00        0.30    0.00      0.00      2.85
#> Dave               0.00     0.0   0.00        0.75    0.15      0.00      2.40
#> Jaxson             0.30     0.0   2.70        0.00    0.00      0.15      2.40
#> Marguerite         0.00     0.0   0.75        0.00    0.00      0.30      5.40
#> Janet              0.00     0.0   0.00        0.00    0.00      0.00      1.05
#> Arya               0.45     0.0   0.00        0.00    0.00      0.75      4.50
#> Priscilla          0.00     0.0   0.00        0.00    0.00      0.00      4.50
#> Essie              0.00     0.0   0.00        0.00    0.00      0.00      3.15
#> Grayson            0.30     0.0   0.45        0.00    0.00      0.00      6.60
#> Roberto            5.40     1.2   5.25        3.90    3.90      5.55      0.00
(triad_cen <- node_triad_census(task_eg))
#>            003 012 102 021D 021U 021C 111D 111U 030T 030C 201 120D 120U 120C
#> Goldie      36   9  45    0    0    0    0    1    0    0   3    0    0    0
#> Troy        45  10  38    0    0    0    1    1    0    0   1    0    0    0
#> Jermaine    55   0  44    0    0    0    0    0    0    0   0    0    0    0
#> Jacob      102   0   3    0    0    0    0    0    0    0   0    0    0    0
#> Leona       55   0  37    0    0    0    2    1    0    0   4    0    0    0
#> Veronica    65  19  10    0    0    0    5    0    0    0   3    1    0    0
#> Latoya      91   1  13    0    0    0    0    0    0    0   0    0    0    0
#> Dave        91   1  13    0    0    0    0    0    0    0   0    0    0    0
#> Jaxson      90   1  14    0    0    0    0    0    0    0   0    0    0    0
#> Marguerite  86   1  18    0    0    0    0    0    0    0   0    0    0    0
#> Janet       95   0  10    0    0    0    0    0    0    0   0    0    0    0
#> Arya        90   1  14    0    0    0    0    0    0    0   0    0    0    0
#> Priscilla   93   1  11    0    0    0    0    0    0    0   0    0    0    0
#> Essie       90   3  12    0    0    0    0    0    0    0   0    0    0    0
#> Grayson     88   2  15    0    0    0    0    0    0    0   0    0    0    0
#> Roberto      0   0   0    0    0    0    0    0    0    0  73    0    0    0
#>            210 300
#> Goldie       4   7
#> Troy         3   6
#> Jermaine     0   6
#> Jacob        0   0
#> Leona        0   6
#> Veronica     2   0
#> Latoya       0   0
#> Dave         0   0
#> Jaxson       0   0
#> Marguerite   0   0
#> Janet        0   0
#> Arya         0   0
#> Priscilla    0   0
#> Essie        0   0
#> Grayson      0   0
#> Roberto      6  26
(quad_cen <- node_quad_census(ison_southern_women))
#>             E4  I41  I40  H4 L42 L41 L40 U42 U41 Y43 Y41  04
#> EVELYN    1402 1046  474 202 397 692 148 200  78  56 190  75
#> LAURA     1539  986  579 202 323 659 182 134  85  35 168  68
#> THERESA   1499  975  434 147 364 784  91 217  77  56 225  91
#> BRENDA    1559  968  562 198 320 682 179 134  82  35 170  71
#> CHARLOTTE 1945  763  972 171 117 419 359  30  89   4  70  21
#> FRANCES   2063  652  944 126 102 502 269  36 102   4 130  30
#> ELEANOR   2122  598  930  98  93 546 224  42 114   4 156  33
#> PEARL     2313  439 1089  65  42 448 252  28 112   1 154  17
#> RUTH      2180  549  918  73  80 577 178  56 123   4 190  32
#> VERNE     2166  561  920  82  83 567 190  53 118   4 184  32
#> MYRA      2111  609  926 116  94 529 239  43 100   4 158  31
#> KATHERINE 1707  897  682 192 239 613 211  97  84  20 164  54
#> SYLVIA    1614  930  544 170 299 718 142 152  82  35 200  74
#> NORA      1387 1058  438 215 403 700 199 209  89  56 146  60
#> HELEN     1503 1021  582 219 327 623 215 144  89  35 148  54
#> DOROTHY   2111  609  926 116  94 529 239  43 100   4 158  31
#> OLIVIA    2341  425 1282  97  18 233 437  10  57   0  58   2
#> FLORA     2341  425 1282  97  18 233 437  10  57   0  58   2
#> E1        2113  629 1165 108  64 325 376   8 105   1  51  15
#> E2        2123  619 1172  96  63 332 359   8 114   1  57  16
#> E3        1554 1014  772 214 284 528 274  52 116  20  78  54
#> E4        1899  804 1038 147 123 384 339  21 117   4  60  24
#> E5        1216 1154  582 264 487 594 226 104 112  56  84  81
#> E6        1239 1159  602 206 456 597 183 145 146  56 100  71
#> E7         994 1200  432 174 630 666 114 270 156 120 114  90
#> E8         603 1063  170 139 975 724  43 507  85 364 131 156
#> E9         671 1210  344 259 884 573 125 344 137 220 101  92
#> E10       1520 1043  792 224 291 498 288  48 118  20  67  51
#> E11       1767  936 1094 207 131 244 415  29  89   4  36   8
#> E12       1343 1126  695 252 391 533 262  70 119  35  70  64
#> E13       1924  777 1014 148 124 414 338  16 112   4  61  28
#> E14       1924  777 1014 148 124 414 338  16 112   4  61  28
group_tie_census(task_eg, cutree(cluster_structural_equivalence(task_eg), 4))
#>         fromGoldie fromTroy fromJermaine fromJacob fromLeona fromVeronica
#> Block 1       0.08      0.2         0.00      0.01      0.04         0.01
#> Block 2       0.15      0.0         0.07      0.00      0.07         0.60
#> Block 3       0.00      0.0         0.45      0.00      1.43         0.15
#> Block 4       5.10      4.8         0.30      0.90      1.80         0.90
#>         fromLatoya fromDave fromJaxson fromMarguerite fromJanet fromArya
#> Block 1       0.15     0.15        0.3           0.16      0.00     0.33
#> Block 2       0.00     0.00        0.0           0.00      0.98     0.00
#> Block 3       0.00     0.00        0.0           0.15      0.22     0.00
#> Block 4       2.85     2.40        2.4           5.40      1.05     4.50
#>         fromPriscilla fromEssie fromGrayson fromRoberto toGoldie toTroy
#> Block 1          0.18      0.07         0.1        3.79     0.04   0.27
#> Block 2          0.00      0.07         0.0        1.57     0.15   0.00
#> Block 3          0.00      0.07         0.0        1.20     0.00   0.00
#> Block 4          4.50      3.15         6.6        0.00     4.65   4.80
#>         toJermaine toJacob toLeona toVeronica toLatoya toDave toJaxson
#> Block 1       0.00    0.01    0.04       0.04     0.14   0.12     0.30
#> Block 2       0.07    0.00    0.07       0.68     0.00   0.00     0.00
#> Block 3       0.45    0.00    1.12       0.22     0.00   0.00     0.00
#> Block 4       0.75    1.05    2.40       1.20     2.40   1.95     2.85
#>         toMarguerite toJanet toArya toPriscilla toEssie toGrayson toRoberto
#> Block 1         0.14    0.00   0.37        0.18    0.03      0.14      3.87
#> Block 2         0.00    1.20   0.00        0.07    0.00      0.00      1.05
#> Block 3         0.07    0.15   0.00        0.00    0.00      0.00      0.98
#> Block 4         5.40    1.20   5.25        3.90    3.90      5.55      0.00
group_triad_census(task_eg, cutree(cluster_regular_equivalence(task_eg), 4))
#>           003  012   102 021D 021U 021C 111D 111U 030T 030C   201 120D 120U
#> Block 1 36.00 9.00 45.00    0    0    0 0.00 1.00    0    0  3.00 0.00    0
#> Block 2 51.67 3.33 39.67    0    0    0 1.00 0.67    0    0  1.67 0.00    0
#> Block 3 89.18 2.73 12.09    0    0    0 0.45 0.00    0    0  0.27 0.09    0
#> Block 4  0.00 0.00  0.00    0    0    0 0.00 0.00    0    0 73.00 0.00    0
#>         120C  210 300
#> Block 1    0 4.00   7
#> Block 2    0 1.00   6
#> Block 3    0 0.18   0
#> Block 4    0 6.00  26