Census by nodes or clusters

These functions include ways to take a census of the positions of nodes in a network. These include a triad census based on the triad profile of nodes, but also a tie census based on the particular tie partners of nodes. Included also are group census functions for summarising the profiles of clusters of nodes in a network.

node_tie_census(object)

node_triad_census(object)

node_quad_census(object)

group_tie_census(object, clusters, decimals = 2)

group_triad_census(object, clusters, decimals = 2)

Arguments

object

A migraph-consistent object.

clusters

a vector of cluster assignment.

decimals

Number of decimal points to round to.

Examples

task_eg <- to_named(to_uniplex(ison_m182, "task_tie"))
(tie_cen <- node_tie_census(task_eg))
#>          fromBrielle fromAntonio fromMakenzie fromFernando fromJeanne
#> Brielle         0.00        0.00         0.00         0.00       0.30
#> Antonio         0.00        0.00         0.00         0.00       0.00
#> Makenzie        0.00        0.00         0.00         0.00       0.15
#> Fernando        0.00        0.00         0.00         0.00       0.00
#> Jeanne          0.30        0.00         0.15         0.00       0.00
#> Harrison        0.00        0.00         0.60         0.00       0.75
#> Paige           0.00        0.75         0.00         0.00       0.00
#> Lisa            0.00        0.30         0.00         0.00       0.00
#> Autumn          0.30        0.00         0.00         0.00       0.00
#> Arthur          0.30        0.00         0.00         0.15       0.00
#> Scott           0.00        0.00         0.30         0.00       2.10
#> Clifford        0.15        0.00         0.00         0.00       0.00
#> Theodore        0.00        0.90         0.00         0.00       0.15
#> Evan            0.00        0.15         0.00         0.00       0.00
#> Verna           0.15        0.15         0.00         0.00       0.00
#> Alondra         5.10        4.80         0.30         0.90       1.80
#>          fromHarrison fromPaige fromLisa fromAutumn fromArthur fromScott
#> Brielle          0.00      0.00     0.00       0.15       0.15      0.00
#> Antonio          0.00      1.20     0.45       0.00       0.00      0.00
#> Makenzie         0.75      0.00     0.00       0.00       0.00      0.15
#> Fernando         0.00      0.00     0.00       0.00       0.15      0.00
#> Jeanne           0.45      0.00     0.00       0.00       0.00      1.80
#> Harrison         0.00      0.00     0.00       0.00       0.30      0.45
#> Paige            0.00      0.00     0.30       0.00       0.00      0.00
#> Lisa             0.00      0.15     0.00       0.00       0.00      0.00
#> Autumn           0.00      0.00     0.00       0.00       0.45      0.00
#> Arthur           0.15      0.00     0.00       0.30       0.00      0.00
#> Scott            0.30      0.00     0.00       0.00       0.00      0.00
#> Clifford         0.00      0.00     0.00       2.70       0.75      0.00
#> Theodore         0.00      0.30     0.75       0.00       0.00      0.00
#> Evan             0.00      0.00     0.15       0.00       0.00      0.00
#> Verna            0.00      0.00     0.00       0.15       0.30      0.00
#> Alondra          0.90      2.85     2.40       2.40       5.40      1.05
#>          fromClifford fromTheodore fromEvan fromVerna fromAlondra toBrielle
#> Brielle          0.00         0.00     0.00      0.15        4.65      0.00
#> Antonio          0.00         1.05     0.30      0.00        4.80      0.00
#> Makenzie         0.00         0.00     0.00      0.00        0.75      0.00
#> Fernando         0.00         0.00     0.00      0.00        1.05      0.00
#> Jeanne           0.00         0.00     0.15      0.00        2.40      0.30
#> Harrison         0.00         0.00     0.15      0.00        1.20      0.00
#> Paige            0.00         0.30     0.15      0.00        2.40      0.00
#> Lisa             0.00         0.60     0.30      0.00        1.95      0.00
#> Autumn           2.40         0.00     0.00      0.15        2.85      0.15
#> Arthur           0.45         0.00     0.00      0.30        5.40      0.15
#> Scott            0.00         0.00     0.00      0.00        1.20      0.00
#> Clifford         0.00         0.00     0.00      0.45        5.25      0.00
#> Theodore         0.00         0.00     0.00      0.00        3.90      0.00
#> Evan             0.00         0.00     0.00      0.00        3.90      0.00
#> Verna            0.75         0.00     0.00      0.00        5.55      0.15
#> Alondra          4.50         4.50     3.15      6.60        0.00      4.65
#>          toAntonio toMakenzie toFernando toJeanne toHarrison toPaige toLisa
#> Brielle       0.00       0.00       0.00     0.30       0.00    0.00   0.00
#> Antonio       0.00       0.00       0.00     0.00       0.00    0.75   0.30
#> Makenzie      0.00       0.00       0.00     0.15       0.60    0.00   0.00
#> Fernando      0.00       0.00       0.00     0.00       0.00    0.00   0.00
#> Jeanne        0.00       0.15       0.00     0.00       0.75    0.00   0.00
#> Harrison      0.00       0.75       0.00     0.45       0.00    0.00   0.00
#> Paige         1.20       0.00       0.00     0.00       0.00    0.00   0.15
#> Lisa          0.45       0.00       0.00     0.00       0.00    0.30   0.00
#> Autumn        0.00       0.00       0.00     0.00       0.00    0.00   0.00
#> Arthur        0.00       0.00       0.15     0.00       0.30    0.00   0.00
#> Scott         0.00       0.15       0.00     1.80       0.45    0.00   0.00
#> Clifford      0.00       0.00       0.00     0.00       0.00    0.00   0.00
#> Theodore      1.05       0.00       0.00     0.00       0.00    0.30   0.60
#> Evan          0.30       0.00       0.00     0.15       0.15    0.15   0.30
#> Verna         0.00       0.00       0.00     0.00       0.00    0.00   0.00
#> Alondra       4.80       0.75       1.05     2.40       1.20    2.40   1.95
#>          toAutumn toArthur toScott toClifford toTheodore toEvan toVerna
#> Brielle      0.30     0.30     0.0       0.15       0.00   0.00    0.15
#> Antonio      0.00     0.00     0.0       0.00       0.90   0.15    0.15
#> Makenzie     0.00     0.00     0.3       0.00       0.00   0.00    0.00
#> Fernando     0.00     0.15     0.0       0.00       0.00   0.00    0.00
#> Jeanne       0.00     0.00     2.1       0.00       0.15   0.00    0.00
#> Harrison     0.00     0.15     0.3       0.00       0.00   0.00    0.00
#> Paige        0.00     0.00     0.0       0.00       0.30   0.00    0.00
#> Lisa         0.00     0.00     0.0       0.00       0.75   0.15    0.00
#> Autumn       0.00     0.30     0.0       2.70       0.00   0.00    0.15
#> Arthur       0.45     0.00     0.0       0.75       0.00   0.00    0.30
#> Scott        0.00     0.00     0.0       0.00       0.00   0.00    0.00
#> Clifford     2.40     0.45     0.0       0.00       0.00   0.00    0.75
#> Theodore     0.00     0.00     0.0       0.00       0.00   0.00    0.00
#> Evan         0.00     0.00     0.0       0.00       0.00   0.00    0.00
#> Verna        0.15     0.30     0.0       0.45       0.00   0.00    0.00
#> Alondra      2.85     5.40     1.2       5.25       3.90   3.90    5.55
#>          toAlondra
#> Brielle       5.10
#> Antonio       4.80
#> Makenzie      0.30
#> Fernando      0.90
#> Jeanne        1.80
#> Harrison      0.90
#> Paige         2.85
#> Lisa          2.40
#> Autumn        2.40
#> Arthur        5.40
#> Scott         1.05
#> Clifford      4.50
#> Theodore      4.50
#> Evan          3.15
#> Verna         6.60
#> Alondra       0.00
(triad_cen <- node_triad_census(task_eg))
#>          003 012 102 021D 021U 021C 111D 111U 030T 030C 201 120D 120U 120C 210
#> Brielle   36   9  45    0    0    0    0    1    0    0   3    0    0    0   4
#> Antonio   45  10  38    0    0    0    1    1    0    0   1    0    0    0   3
#> Makenzie  55   0  44    0    0    0    0    0    0    0   0    0    0    0   0
#> Fernando 102   0   3    0    0    0    0    0    0    0   0    0    0    0   0
#> Jeanne    55   0  37    0    0    0    2    1    0    0   4    0    0    0   0
#> Harrison  65  19  10    0    0    0    5    0    0    0   3    1    0    0   2
#> Paige     91   1  13    0    0    0    0    0    0    0   0    0    0    0   0
#> Lisa      91   1  13    0    0    0    0    0    0    0   0    0    0    0   0
#> Autumn    90   1  14    0    0    0    0    0    0    0   0    0    0    0   0
#> Arthur    86   1  18    0    0    0    0    0    0    0   0    0    0    0   0
#> Scott     95   0  10    0    0    0    0    0    0    0   0    0    0    0   0
#> Clifford  90   1  14    0    0    0    0    0    0    0   0    0    0    0   0
#> Theodore  93   1  11    0    0    0    0    0    0    0   0    0    0    0   0
#> Evan      90   3  12    0    0    0    0    0    0    0   0    0    0    0   0
#> Verna     88   2  15    0    0    0    0    0    0    0   0    0    0    0   0
#> Alondra    0   0   0    0    0    0    0    0    0    0  73    0    0    0   6
#>          300
#> Brielle    7
#> Antonio    6
#> Makenzie   6
#> Fernando   0
#> Jeanne     6
#> Harrison   0
#> Paige      0
#> Lisa       0
#> Autumn     0
#> Arthur     0
#> Scott      0
#> Clifford   0
#> Theodore   0
#> Evan       0
#> Verna      0
#> Alondra   26
(quad_cen <- node_quad_census(southern_women))
#>             E4  I41  I40  H4 L42 L41 L40 U42 U41 Y43 Y41  04
#> EVELYN    1402 1046  474 202 397 692 148 200  78  56 190  75
#> LAURA     1539  986  579 202 323 659 182 134  85  35 168  68
#> THERESA   1499  975  434 147 364 784  91 217  77  56 225  91
#> BRENDA    1559  968  562 198 320 682 179 134  82  35 170  71
#> CHARLOTTE 1945  763  972 171 117 419 359  30  89   4  70  21
#> FRANCES   2063  652  944 126 102 502 269  36 102   4 130  30
#> ELEANOR   2122  598  930  98  93 546 224  42 114   4 156  33
#> PEARL     2313  439 1089  65  42 448 252  28 112   1 154  17
#> RUTH      2180  549  918  73  80 577 178  56 123   4 190  32
#> VERNE     2166  561  920  82  83 567 190  53 118   4 184  32
#> MYRA      2111  609  926 116  94 529 239  43 100   4 158  31
#> KATHERINE 1707  897  682 192 239 613 211  97  84  20 164  54
#> SYLVIA    1614  930  544 170 299 718 142 152  82  35 200  74
#> NORA      1387 1058  438 215 403 700 199 209  89  56 146  60
#> HELEN     1503 1021  582 219 327 623 215 144  89  35 148  54
#> DOROTHY   2111  609  926 116  94 529 239  43 100   4 158  31
#> OLIVIA    2341  425 1282  97  18 233 437  10  57   0  58   2
#> FLORA     2341  425 1282  97  18 233 437  10  57   0  58   2
#> E1        2113  629 1165 108  64 325 376   8 105   1  51  15
#> E2        2123  619 1172  96  63 332 359   8 114   1  57  16
#> E3        1554 1014  772 214 284 528 274  52 116  20  78  54
#> E4        1899  804 1038 147 123 384 339  21 117   4  60  24
#> E5        1216 1154  582 264 487 594 226 104 112  56  84  81
#> E6        1239 1159  602 206 456 597 183 145 146  56 100  71
#> E7         994 1200  432 174 630 666 114 270 156 120 114  90
#> E8         603 1063  170 139 975 724  43 507  85 364 131 156
#> E9         671 1210  344 259 884 573 125 344 137 220 101  92
#> E10       1520 1043  792 224 291 498 288  48 118  20  67  51
#> E11       1767  936 1094 207 131 244 415  29  89   4  36   8
#> E12       1343 1126  695 252 391 533 262  70 119  35  70  64
#> E13       1924  777 1014 148 124 414 338  16 112   4  61  28
#> E14       1924  777 1014 148 124 414 338  16 112   4  61  28
group_tie_census(task_eg, cutree(cluster_structural_equivalence(task_eg), 4))
#>         fromBrielle fromAntonio fromMakenzie fromFernando fromJeanne
#> Block 1        0.08         0.2         0.00         0.01       0.04
#> Block 2        0.15         0.0         0.07         0.00       0.07
#> Block 3        0.00         0.0         0.45         0.00       1.43
#> Block 4        5.10         4.8         0.30         0.90       1.80
#>         fromHarrison fromPaige fromLisa fromAutumn fromArthur fromScott
#> Block 1         0.01      0.15     0.15        0.3       0.16      0.00
#> Block 2         0.60      0.00     0.00        0.0       0.00      0.98
#> Block 3         0.15      0.00     0.00        0.0       0.15      0.22
#> Block 4         0.90      2.85     2.40        2.4       5.40      1.05
#>         fromClifford fromTheodore fromEvan fromVerna fromAlondra toBrielle
#> Block 1         0.33         0.18     0.07       0.1        3.79      0.04
#> Block 2         0.00         0.00     0.07       0.0        1.57      0.15
#> Block 3         0.00         0.00     0.07       0.0        1.20      0.00
#> Block 4         4.50         4.50     3.15       6.6        0.00      4.65
#>         toAntonio toMakenzie toFernando toJeanne toHarrison toPaige toLisa
#> Block 1      0.27       0.00       0.01     0.04       0.04    0.14   0.12
#> Block 2      0.00       0.07       0.00     0.07       0.68    0.00   0.00
#> Block 3      0.00       0.45       0.00     1.12       0.22    0.00   0.00
#> Block 4      4.80       0.75       1.05     2.40       1.20    2.40   1.95
#>         toAutumn toArthur toScott toClifford toTheodore toEvan toVerna
#> Block 1     0.30     0.14    0.00       0.37       0.18   0.03    0.14
#> Block 2     0.00     0.00    1.20       0.00       0.07   0.00    0.00
#> Block 3     0.00     0.07    0.15       0.00       0.00   0.00    0.00
#> Block 4     2.85     5.40    1.20       5.25       3.90   3.90    5.55
#>         toAlondra
#> Block 1      3.87
#> Block 2      1.05
#> Block 3      0.98
#> Block 4      0.00
group_triad_census(task_eg, cutree(cluster_regular_equivalence(task_eg), 4))
#>           003  012   102 021D 021U 021C 111D 111U 030T 030C   201 120D 120U
#> Block 1 36.00 9.00 45.00    0    0    0 0.00 1.00    0    0  3.00 0.00    0
#> Block 2 51.67 3.33 39.67    0    0    0 1.00 0.67    0    0  1.67 0.00    0
#> Block 3 89.18 2.73 12.09    0    0    0 0.45 0.00    0    0  0.27 0.09    0
#> Block 4  0.00 0.00  0.00    0    0    0 0.00 0.00    0    0 73.00 0.00    0
#>         120C  210 300
#> Block 1    0 4.00   7
#> Block 2    0 1.00   6
#> Block 3    0 0.18   0
#> Block 4    0 6.00  26