Centrality for one- and two-mode networks

These functions calculate common centrality measures for both one- and two-mode networks. They accept as objects matrices and igraph graphs, and can be used within a tidygraph workflow. Importantly, these functions also offer correct normalization for two-mode networks.

node_degree(
  object,
  weights = NULL,
  mode = "out",
  loops = TRUE,
  normalized = FALSE
)

node_closeness(
  object,
  weights = NULL,
  mode = "out",
  normalized = FALSE,
  cutoff = NULL
)

node_betweenness(
  object,
  weights = NULL,
  directed = TRUE,
  cutoff = NULL,
  nobigint = TRUE,
  normalized = FALSE
)

node_eigenvector(
  object,
  weights = NULL,
  directed = FALSE,
  options = igraph::arpack_defaults,
  scale = FALSE,
  normalized = FALSE
)

Arguments

object

Either an igraph graph object or a matrix.

weights

The weight of the edges to use for the calculation. Will be evaluated in the context of the edge data.

mode

How should edges be followed. Ignored for undirected graphs

loops

Should loops be included in the calculation

normalized

For one-mode networks, should Borgatti and Everett normalization be applied?

cutoff

maximum path length to use during calculations

directed

Should direction of edges be used for the calculations

nobigint

Should big integers be avoided during calculations

options

Settings passed on to igraph::arpack()

scale

Should the scores be scaled to range between 0 and 1?

Value

Depending on how and what kind of an object is passed to the function, the function will return a tidygraph object where the nodes have been updated A numeric vector giving the betweenness centrality measure of each node. A numeric vector giving the eigenvector centrality measure of each node.

References

Borgatti, Stephen P., and Martin G. Everett. "Network analysis of 2-mode data." Social networks 19.3 (1997): 243-270.

Faust, Katherine. "Centrality in affiliation networks." Social networks 19.2 (1997): 157-191.

See also

Examples

node_degree(mpn_elite_mex)
#>   Alvarezx     Beteta     Blanco Bustamente    Gortari     Lozano     Madrid 
#>          2          5          2          3          5          6          5 
#>    Margain       Mena   Portillo      Rojas 
#>          8          3          4          1 
node_degree(southern_women)
#>    EVELYN     LAURA   THERESA    BRENDA CHARLOTTE   FRANCES   ELEANOR     PEARL 
#>         8         7         8         7         4         4         4         3 
#>      RUTH     VERNE      MYRA KATHERINE    SYLVIA      NORA     HELEN   DOROTHY 
#>         4         4         4         6         7         8         7         4 
#>    OLIVIA     FLORA        E1        E2        E3        E4        E5        E6 
#>         2         2         3         3         6         4         8         8 
#>        E7        E8        E9       E10       E11       E12       E13       E14 
#>        10        14        12         6         4         7         4         4 
node_closeness(mpn_elite_mex)
#>   Alvarezx     Beteta     Blanco Bustamente    Gortari     Lozano     Madrid 
#> 0.04761905 0.06666667 0.05263158 0.05555556 0.06250000 0.07142857 0.06250000 
#>    Margain       Mena   Portillo      Rojas 
#> 0.07692308 0.05555556 0.05882353 0.03703704 
node_closeness(southern_women)
#>     EVELYN      LAURA    THERESA     BRENDA  CHARLOTTE    FRANCES    ELEANOR 
#> 0.01666667 0.01515152 0.01666667 0.01515152 0.01250000 0.01388889 0.01388889 
#>      PEARL       RUTH      VERNE       MYRA  KATHERINE     SYLVIA       NORA 
#> 0.01388889 0.01470588 0.01470588 0.01428571 0.01515152 0.01612903 0.01666667 
#>      HELEN    DOROTHY     OLIVIA      FLORA         E1         E2         E3 
#> 0.01612903 0.01428571 0.01219512 0.01219512 0.01190476 0.01190476 0.01282051 
#>         E4         E5         E6         E7         E8         E9        E10 
#> 0.01219512 0.01351351 0.01562500 0.01666667 0.01923077 0.01785714 0.01282051 
#>        E11        E12        E13        E14 
#> 0.01219512 0.01315789 0.01219512 0.01219512 
node_betweenness(mpn_elite_mex)
#>   Alvarezx     Beteta     Blanco Bustamente    Gortari     Lozano     Madrid 
#>   0.000000  13.333333   0.000000  18.500000   2.166667  19.000000   2.166667 
#>    Margain       Mena   Portillo      Rojas 
#>  25.500000   0.000000   3.333333   0.000000 
node_betweenness(southern_women)
#>      EVELYN       LAURA     THERESA      BRENDA   CHARLOTTE     FRANCES 
#>  42.7599983  22.8565396  38.7392641  22.0119104   4.7279418   4.7516238 
#>     ELEANOR       PEARL        RUTH       VERNE        MYRA   KATHERINE 
#>   4.1356871   2.9762664   7.3608922   6.3676024   5.9435242  16.2889210 
#>      SYLVIA        NORA       HELEN     DOROTHY      OLIVIA       FLORA 
#>  25.2987307  43.9377835  30.7265229   5.9435242   2.0866336   2.0866336 
#>          E1          E2          E3          E4          E5          E6 
#>   0.9737485   0.9440910   8.1977508   3.4529893  16.9812110  28.0103112 
#>          E7          E8          E9         E10         E11         E12 
#>  58.0969236 108.2616581  96.2294880   6.8185655   9.0194398  10.2353650 
#>         E13         E14 
#>   1.8892292   1.8892292 
node_eigenvector(mpn_elite_mex)
#>   Alvarezx     Beteta     Blanco Bustamente    Gortari     Lozano     Madrid 
#> 0.15638683 0.34382531 0.17579240 0.15407997 0.37681994 0.37444178 0.38138413 
#>    Margain       Mena   Portillo      Rojas 
#> 0.48115951 0.25321407 0.27999270 0.03165737 
node_eigenvector(southern_women)
#>     EVELYN      LAURA    THERESA     BRENDA  CHARLOTTE    FRANCES    ELEANOR 
#> 0.29913625 0.28092924 0.33382479 0.28434268 0.16055764 0.20319443 0.22547488 
#>      PEARL       RUTH      VERNE       MYRA  KATHERINE     SYLVIA       NORA 
#> 0.18677117 0.23829269 0.23101265 0.20662026 0.24263486 0.29098916 0.27036422 
#>      HELEN    DOROTHY     OLIVIA      FLORA         E1         E2         E3 
#> 0.24964055 0.20662026 0.07752323 0.07752323 0.15211097 0.16106257 0.25183346 
#>         E4         E5         E6         E7         E8         E9        E10 
#> 0.18428786 0.30462311 0.31606479 0.36943483 0.45192837 0.35686135 0.22825957 
#>        E11        E12        E13        E14 
#> 0.11234306 0.25511942 0.17743466 0.17743466