Measures of network cohesion or connectedness

These functions return values or vectors relating to how connected a network is and the number of nodes or edges to remove that would increase fragmentation.

network_density(.data)

network_components(.data)

network_cohesion(.data)

network_adhesion(.data)

network_diameter(.data)

network_length(.data)

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

Functions

network_density(): Summarises the ratio of ties to the number of possible ties.

network_components(): Returns number of (strong) components in the network. To get the 'weak' components of a directed graph, please use manynet::to_undirected() first.

network_cohesion(): Returns the minimum number of nodes to remove from the network needed to increase the number of components.

network_adhesion(): Returns the minimum number of edges needed to remove from the network to increase the number of components.

network_diameter(): Returns the maximum path length in the network.

network_length(): Returns the average path length in the network.

References

White, Douglas R and Frank Harary. 2001. "The Cohesiveness of Blocks In Social Networks: Node Connectivity and Conditional Density." Sociological Methodology 31(1): 305-59.

See also

Examples

network_density(mpn_elite_mex)
#> [1] 0.197
network_density(mpn_elite_usa_advice)
#> [1] 0.164
network_components(mpn_ryanair)
#> [1] 2
network_components(manynet::to_undirected(mpn_ryanair))
#> [1] 1
network_cohesion(manynet::ison_marvel_relationships)
#> [1] 0
network_cohesion(manynet::to_giant(manynet::ison_marvel_relationships))
#> [1] 5
network_adhesion(manynet::ison_marvel_relationships)
#> [1] 0
network_adhesion(manynet::to_giant(manynet::ison_marvel_relationships))
#> [1] 5
network_diameter(manynet::ison_marvel_relationships)
#> [1] 3
network_diameter(manynet::to_giant(manynet::ison_marvel_relationships))
#> [1] 3
network_length(manynet::ison_marvel_relationships)
#> [1] 1.62
network_length(manynet::to_giant(manynet::ison_marvel_relationships))
#> [1] 1.62