Community graph partitioning algorithms

Community graph partitioning algorithms

node_kernighanlin(object)

node_walktrap(object, times = 50)

node_edge_betweenness(object)

node_fast_greedy(object)

Arguments

object

An object of a migraph-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

times

Integer indicating number of simulations/walks used. By default, times=50.

Functions

node_kernighanlin(): A greedy, iterative, deterministic graph partitioning algorithm that results in a graph with two equally-sized communities

node_walktrap(): A hierarchical, agglomerative algorithm based on random walks.

node_edge_betweenness(): A hierarchical, decomposition algorithm where edges are removed in decreasing order of the number of shortest paths passing through the edge, resulting in a hierarchical representation of group membership.

node_fast_greedy(): A hierarchical, agglomerative algorithm, that tries to optimize modularity in a greedy manner.

Walktrap

The general idea is that random walks on a network are more likely to stay within the same community because few edges lead outside a community. By repeating random walks of 4 steps many times, information about the hierarchical merging of communities is collected.

Edge-betweenness

This is motivated by the idea that edges connecting different groups are more likely to lie on multiple shortest paths when they are the only option to go from one group to another. This method yields good results but is very slow because of the computational complexity of edge-betweenness calculations and the betweenness scores have to be re-calculated after every edge removal. Networks of ~700 nodes and ~3500 ties are around the upper size limit that are feasible with this approach.

Fast-greedy

Initially, each node is assigned a separate community. Communities are then merged iteratively such that each merge yields the largest increase in the current value of modularity, until no further increases to the modularity are possible. The method is fast and recommended as a first approximation because it has no parameters to tune. However, it is known to suffer from a resolution limit.

References

Kernighan, Brian W., and Shen Lin. 1970. "An efficient heuristic procedure for partitioning graphs." The Bell System Technical Journal 49(2): 291-307. doi:10.1002/j.1538-7305.1970.tb01770.x

See also

Examples

node_kernighanlin(ison_adolescents)
#> 1 
#>    Betty, Jane, Dale, Pam
#>  2 
#>    Sue, Alice, Carol, Tina
node_kernighanlin(ison_southern_women)
#> 1 
#>    E1, E2, E3, E4, E5, E6, E7, E8, E9, E10, E11, E12, E13, E14 
#>    OLIVIA, FLORA
#>  2 
#>     
#>    EVELYN, LAURA, THERESA, BRENDA, CHARLOTTE, FRANCES, ELEANOR, PEARL, RUTH, VERNE, MYRA, KATHERINE, SYLVIA, NORA, HELEN, DOROTHY
node_walktrap(ison_adolescents)
#> 1 
#>    7, 8
#>  2 
#>    1, 2, 3, 4, 5
#>  3 
#>    6
node_edge_betweenness(ison_adolescents)
#> 1 
#>    1, 2, 3, 4, 5, 6
#>  2 
#>    7, 8
node_fast_greedy(ison_adolescents)
#> 1 
#>    6, 7, 8
#>  2 
#>    3, 4, 5
#>  3 
#>    1, 2