Functions that have been renamed, superseded, or are no longer working

Generally these functions have been superseded or renamed. Upon using them, a message is provided directing the user to the new function. However, at this stage of package development, we generally clear older defunct functions at each minor release, and so you are strongly encouraged to use the new functions/names/syntax wherever possible and update your scripts accordingly.

ggtree(hc, k = NULL)

ggidentify_clusters(hc, census, method = c("elbow", "strict"))

cluster_structural_equivalence(object)

cluster_regular_equivalence(object)

blockmodel_concor(
  object,
  p = 1,
  cutoff = 0.999,
  max.iter = 25,
  block.content = "density"
)

summarise_statistics(node_measure, clusters = NULL, sumFUN = mean)

blockmodel(object, clusters)

# S3 method for block_model
print(x, ...)

reduce_graph(blockmodel, block_labels = NULL)

ggatyear(edgelist, year, ...)

group_tie_census(object, clusters, decimals = 2)

group_triad_census(object, clusters, decimals = 2)

add_edge_attribute(object, attr_name, vector)

edge_attribute(object, attribute)

edge_signs(object)

edge_weights(object)

edge_betweenness(object, normalized = TRUE)

edge_closeness(object, normalized = TRUE)

edge_degree(object, normalized = TRUE)

edge_eigenvector(object, normalized = TRUE)

edge_loop(object)

edge_multiple(object)

edge_bridges(object)

edge_reciprocal(object)

graph_edge_attributes(object)

graph_edges(object)

node_cuts(object)

to_edges(object)

join_edges(object, object2, attr_name)

Functions

ggtree: Deprecated on 2022-05-27.

ggidentify_clusters: Defunct on 2022-05-27.

cluster_structural_equivalence: Deprecated on 2022-05-27.

cluster_regular_equivalence: Deprecated on 2022-05-27.

blockmodel_concor: Deprecated on 2022-05-30.

summarise_statistics: Deprecated on 2022-06-03.

blockmodel: Deprecated on 2022-06-03.

print.block_model: Deprecated on 2022-06-03.

reduce_graph: Deprecated on 2022-06-03.

ggatyear: Deprecated on 2022-06-08.

group_tie_census: Deprecated on 2022-06-09.

group_triad_census: Deprecated on 2022-06-09.

add_edge_attribute: Deprecated on 2022-06-28.

edge_attribute: Deprecated on 2022-06-28.

edge_signs: Deprecated on 2022-06-28.

edge_weights: Deprecated on 2022-06-28.

edge_betweenness: Deprecated on 2022-06-28.

edge_closeness: Deprecated on 2022-06-28.

edge_degree: Deprecated on 2022-06-28.

edge_eigenvector: Deprecated on 2022-06-28.

edge_loop: Deprecated on 2022-06-28.

edge_multiple: Deprecated on 2022-06-28.

edge_bridges: Deprecated on 2022-06-28.

edge_reciprocal: Deprecated on 2022-06-28.

graph_edge_attributes: Deprecated on 2022-06-28.

graph_edges: Deprecated on 2022-06-28.

node_cuts: Deprecated on 2022-06-30.

to_edges: Deprecated on 2022-06-28.

join_edges: Deprecated on 2022-07-03.