Functions that have been renamed, superseded, or are no longer working

Generally these functions have been superseded or renamed. Upon using them, a message is provided directing the user to the new function. However, at this stage of package development, we generally clear older defunct functions at each minor release, and so you are strongly encouraged to use the new functions/names/syntax wherever possible and update your scripts accordingly.

edge_betweenness(object, normalized = TRUE)

edge_closeness(object, normalized = TRUE)

edge_degree(object, normalized = TRUE)

edge_eigenvector(object, normalized = TRUE)

edge_loop(object)

edge_multiple(object)

edge_bridges(object)

edge_reciprocal(object)

node_cuts(object)

graph_blau_index(object, attribute, clusters = NULL)

graph_diversity(object, attribute, clusters = NULL)

graph_ei_index(object, attribute)

graph_homophily(object, attribute)

# S3 method for graph_test
plot(x, ..., threshold = 0.95, tails = c("two", "one"))

# S3 method for graph_test
print(x, ..., max.length = 6, digits = 3)

# S3 method for graph_measure
print(x, ..., digits = 3)

# S3 method for graph_motif
print(x, ...)

graph_adhesion(object)

graph_cohesion(object)

graph_assortativity(object)

graph_balance(object)

graph_betweenness(object, normalized = TRUE, direction = c("all", "out", "in"))

graph_closeness(object, normalized = TRUE, direction = c("all", "out", "in"))

graph_degree(object, normalized = TRUE, direction = c("all", "out", "in"))

graph_eigenvector(object, normalized = TRUE)

graph_components(object)

graph_factions(object, membership = NULL)

graph_congruency(object, object2)

graph_equivalency(object)

graph_core(object, membership = NULL)

graph_density(object)

graph_reciprocity(object, method = "default")

graph_transitivity(object)

graph_diameter(object)

graph_length(object)

graph_dyad_census(object)

graph_triad_census(object)

graph_mixed_census(object, object2)

graph_modularity(object, membership = NULL, resolution = 1)

graph_smallworld(object, times = 100)

network_homophily(object, attribute)

node_homophily(object, attribute)

Functions

edge_betweenness(): Deprecated on 2022-06-28.

edge_closeness(): Deprecated on 2022-06-28.

edge_degree(): Deprecated on 2022-06-28.

edge_eigenvector(): Deprecated on 2022-06-28.

edge_loop(): Deprecated on 2022-06-28.

edge_multiple(): Deprecated on 2022-06-28.

edge_bridges(): Deprecated on 2022-06-28.

edge_reciprocal(): Deprecated on 2022-06-28.

node_cuts(): Deprecated on 2022-06-30.

graph_blau_index(): Deprecated on 2022-09-10.

graph_diversity(): Deprecated on 2022-09-25.

graph_ei_index(): Deprecated on 2022-09-10.

graph_homophily(): Deprecated on 2022-09-25.

plot(graph_test): Deprecated on 2022-09-25.

print(graph_test): Deprecated on 2022-09-25.

print(graph_measure): Deprecated on 2022-09-25.

print(graph_motif): Deprecated on 2022-09-25.

graph_adhesion(): Deprecated on 2022-09-25.

graph_cohesion(): Deprecated on 2022-09-25.

graph_assortativity(): Deprecated on 2022-09-25.

graph_balance(): Deprecated on 2022-09-25.

graph_betweenness(): Deprecated on 2022-09-25.

graph_closeness(): Deprecated on 2022-09-25.

graph_degree(): Deprecated on 2022-09-25.

graph_eigenvector(): Deprecated on 2022-09-25.

graph_components(): Deprecated on 2022-09-25.

graph_factions(): Deprecated on 2022-09-25.

graph_congruency(): Deprecated on 2022-09-25.

graph_equivalency(): Deprecated on 2022-09-25.

graph_core(): Deprecated on 2022-09-25.

graph_density(): Deprecated on 2022-09-25.

graph_reciprocity(): Deprecated on 2022-09-25.

graph_transitivity(): Deprecated on 2022-09-25.

graph_diameter(): Deprecated on 2022-09-25.

graph_length(): Deprecated on 2022-09-25.

graph_dyad_census(): Deprecated on 2022-09-25.

graph_triad_census(): Deprecated on 2022-09-25.

graph_mixed_census(): Deprecated on 2022-09-25.

graph_modularity(): Deprecated on 2022-09-25.

graph_smallworld(): Deprecated on 2022-09-25.

network_homophily(): Deprecated on 2022-09-25.

node_homophily(): Deprecated on 2022-09-25.