Measures of network topological features

Measures of network topological features

graph_core(object, membership = NULL)

graph_factions(object, membership = NULL)

graph_modularity(object, membership = NULL, resolution = 1)

graph_smallworld(object, times = 100)

graph_balance(object)

Source

{signnet} by David Schoch

Arguments

object

An object of a migraph-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

membership

A vector of partition membership.

resolution

A proportion indicating the resolution scale. By default 1.

times

Integer of number of simulations.

Functions

graph_core(): Returns correlation between a given network and a core-periphery model with the same dimensions.

graph_factions(): Returns correlation between a given network and a component model with the same dimensions.

graph_modularity(): Returns modularity of one- or two-mode networks based on nodes' membership in pre-defined clusters.

graph_smallworld(): Returns small-world metrics for one- and two-mode networks. Small-world networks can be highly clustered and yet have short path lengths.

graph_balance(): Returns the structural balance index on the proportion of balanced triangles, ranging between 0 if all triangles are imbalanced and 1 if all triangles are balanced.

References

Borgatti, Stephen P., and Martin G. Everett. 2000. “Models of Core/Periphery Structures.” Social Networks 21(4):375–95. doi:10.1016/S0378-8733(99)00019-2

Murata, Tsuyoshi. 2010. Modularity for Bipartite Networks. In: Memon, N., Xu, J., Hicks, D., Chen, H. (eds) Data Mining for Social Network Data. Annals of Information Systems, V1ol 12. Springer, Boston, MA. doi:10.1007/978-1-4419-6287-4_7

Watts, Duncan J., and Steven H. Strogatz. 1998. “Collective Dynamics of ‘Small-World’ Networks.” Nature 393(6684):440–42. doi:10.1038/30918 .

See also

Examples

graph_core(ison_adolescents)
#> [1] 0.323
graph_core(ison_southern_women)
#> [1] 0.343
graph_factions(ison_adolescents)
#> [1] 0.174
graph_factions(ison_southern_women)
#> [1] 0.485
graph_modularity(ison_adolescents, 
  node_kernighanlin(ison_adolescents))
#> [1] -0.205
graph_modularity(ison_southern_women, 
  node_kernighanlin(ison_southern_women))
#> [1] -0.458
graph_smallworld(ison_brandes)
#> [1] 0
graph_smallworld(ison_southern_women)
#> [1] 1.32
graph_balance(ison_marvel_relationships)
#> [1] 0.668