Measures of structural holes

These function provide different measures of the degree to which nodes fill structural holes, as outlined in Burt (1992). Burt's theory holds that while those nodes embedded in dense clusters of close connections are likely exposed to the same or similar ideas and information, those who fill structural holes between two otherwise disconnected groups can gain some comparative advantage from that position.

node_bridges(object)

node_redundancy(object)

node_effsize(object)

node_efficiency(object)

node_constraint(object)

node_hierarchy(object)

Arguments

object

An object of a migraph-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

Details

A number of different ways of measuring these structural holes are available. Note that we use Borgatti's reformulation for unweighted networks in node_redundancy() and node_effsize(). Redundancy is thus \(\frac{2t}{n}\), where \(t\) is the sum of ties and \(n\) the sum of nodes in each node's neighbourhood, and effective size is calculated as \(n - \frac{2t}{n}\). Node efficiency is the node's effective size divided by its degree.

Functions

node_bridges(): Returns the sum of bridges to which each node is adjacent.

node_redundancy(): Returns a measure of the redundancy of each nodes' contacts.

node_effsize(): Returns nodes' effective size

node_efficiency(): Returns nodes' efficiency

node_constraint(): Returns nodes' constraint scores for one-mode networks according to Burt (1992) and for two-mode networks according to Hollway et al (2020).

node_hierarchy(): Returns nodes' exposure to hierarchy, where only one or two contacts are the source of closure

References

Burt, Ronald S. 1992. Structural Holes: The Social Structure of Competition. Cambridge, MA: Harvard University Press.

Borgatti, Steven. 1997. “Structural Holes: Unpacking Burt’s Redundancy Measures” Connections 20(1):35-38.

Hollway, James, Jean-Frédéric Morin, and Joost Pauwelyn. 2020. "Structural conditions for novelty: the introduction of new environmental clauses to the trade regime complex." International Environmental Agreements: Politics, Law and Economics 20 (1): 61–83. doi:10.1007/s10784-019-09464-5 .

See also

Examples

node_bridges(ison_adolescents)
#>   Betty   Sue Alice  Jane  Dale   Pam Carol  Tina
#> 1     1     1     0     0     0     1     2     1
node_bridges(ison_southern_women)
#>   EVELYN LAURA THERESA BRENDA CHARLOTTE FRANCES ELEANOR PEARL  RUTH VERNE  MYRA
#> 1      0     0       0      0         0       0       0     0     0     0     0
#> # ... with 7 more from this nodeset in the vector.
#>      E1    E2    E3    E4    E5    E6    E7    E8    E9   E10   E11   E12   E13
#> 1     0     0     0     0     0     0     0     0     0     0     0     0     0
#> # ... with 1 more from this nodeset in the vector.
node_redundancy(ison_adolescents)
#>   Betty   Sue Alice  Jane  Dale   Pam Carol  Tina
#> 1     6   2.5   2.5     4  3.33     4     2     2
node_redundancy(ison_southern_women)
#>   EVELYN LAURA THERESA BRENDA CHARLOTTE FRANCES ELEANOR PEARL  RUTH VERNE  MYRA
#> 1   12.5  12.9    14.2   13.1        12      16      18  20.7    20  19.5  17.5
#> # ... with 7 more from this nodeset in the vector.
#>      E1    E2    E3    E4    E5    E6    E7    E8    E9   E10   E11   E12   E13
#> 1  12.7  13.3  10.7  11.5   9.5  10.2    10     9     8    10   7.5  9.43    12
#> # ... with 1 more from this nodeset in the vector.
node_effsize(ison_adolescents)
#>   Betty   Sue Alice  Jane   Dale   Pam Carol  Tina
#> 1    -5   1.5   1.5    -2 -0.333    -1     0    -1
node_effsize(ison_southern_women)
#>   EVELYN LAURA THERESA BRENDA CHARLOTTE FRANCES ELEANOR PEARL  RUTH VERNE  MYRA
#> 1   -4.5 -5.86   -6.25  -6.14        -8     -12     -14 -17.7   -16 -15.5 -13.5
#> # ... with 7 more from this nodeset in the vector.
#>      E1    E2    E3    E4    E5    E6    E7    E8    E9   E10   E11   E12   E13
#> 1 -9.67 -10.3 -4.67  -7.5  -1.5 -2.25     0     5     4    -4  -3.5 -2.43    -8
#> # ... with 1 more from this nodeset in the vector.
node_efficiency(ison_adolescents)
#>      V1    V2    V3    V4     V5     V6    V7    V8
#> 1    -5 0.375 0.375    -1 -0.111 -0.333     0    -1
node_efficiency(ison_southern_women)
#>       V1     V2     V3     V4    V5    V6    V7    V8    V9   V10   V11   V12
#> 1 -0.562 -0.837 -0.781 -0.878    -2    -3  -3.5 -5.89    -4 -3.88 -3.38 -1.28
#> # ... with 6 more from this nodeset in the vector.
#>      V1    V2     V3    V4     V5     V6    V7    V8    V9    V10    V11    V12
#> 1 -3.22 -3.44 -0.778 -1.88 -0.188 -0.281     0 0.357 0.333 -0.667 -0.875 -0.347
#> # ... with 2 more from this nodeset in the vector.
node_constraint(ison_southern_women)
#>   EVELYN LAURA THERESA BRENDA CHARLOTTE FRANCES ELEANOR PEARL  RUTH VERNE  MYRA
#> 1  0.276 0.306   0.294  0.317     0.384   0.412   0.415 0.461 0.398 0.407 0.434
#> # ... with 7 more from this nodeset in the vector.
#>      E1    E2    E3    E4    E5    E6    E7    E8    E9   E10   E11   E12   E13
#> 1 0.476 0.476 0.338 0.394 0.275 0.241 0.192 0.170 0.161 0.327 0.308 0.292 0.432
#> # ... with 1 more from this nodeset in the vector.
node_hierarchy(ison_adolescents)
#>   Betty    Sue  Alice    Jane   Dale     Pam Carol  Tina
#> 1     0 0.0306 0.0274 0.00587 0.0333 0.00612 0.102     0
node_hierarchy(ison_southern_women)
#>   EVELYN  LAURA THERESA BRENDA CHARLOTTE FRANCES ELEANOR  PEARL   RUTH  VERNE
#> 1 0.0341 0.0359  0.0334 0.0301    0.0234  0.0207  0.0124 0.0154 0.0172 0.0221
#> # ... with 8 more from this nodeset in the vector.
#>        E1       E2      E3      E4      E5     E6      E7      E8     E9    E10
#> 1 0.00150 0.000819 0.00602 0.00576 0.00572 0.0101 0.00573 0.00565 0.0150 0.0111
#> # ... with 4 more from this nodeset in the vector.