Two-mode and three-mode American power elite database (Domhoff 2016)

mpn_elite_usa_advice is a 2-mode network of persons serving as directors or trustees of think tanks. Think tanks are “public-policy research analysis and engagement organizations that generate policy-oriented research, analysis, and advice on domestic and international issues, thereby enabling policymakers and the public to make informed decisions about public policy” (McGann 2016: 6). The Power Elite Database (Domhoff 2016) includes information on the directors of 33 prominent think tanks in 2012. Here we include only 14 directors who held three or more seats among 20 think tanks.

mpn_elite_usa_money is based on 26 elites who sat on the boards of directors for at least two of six economic policy making organizations (Domhoff 2016), and also made campaign contributions to one or more of six candidates running in the primary election contests for the 2008 Presidential nominations of the Republican Party (Rudy Giuliani, John McCain, Mitt Romney) or the Democratic Party (Hillary Clinton, Christopher Dodd, Barack Obama).

data(mpn_elite_usa_advice)

data(mpn_elite_usa_money)

Format

mpn_elite_usa_advice is a two-mode, named, and unweighted tbl_graph with 32 nodes and 46 edges.

mpn_elite_usa_advice is a two-mode, named, and unweighted tbl_graph with 38 nodes and 103 edges.

References

Domhoff, G William. 2016. “Who Rules America? Power Elite Database.”

The Center for Responsive Politics. 2019. “OpenSecrets.”

Knoke, David, Mario Diani, James Hollway, and Dimitris C Christopoulos. 2021. Multimodal Political Networks. Cambridge University Press. Cambridge University Press.