Two-mode 112th Congress Senate Voting (Knoke et al. 2021)

These datasets list the U.S. Senators who served in the 112th Congress, which met from January 3, 2011 to January 3, 2013. Although the Senate has 100 seats, 103 persons served during this period due to two resignations and a death. However, the third replacement occurred only two days before the end and cast no votes on the bills investigated here. Hence, the number of Senators analyzed is 102.

CQ Almanac identified 25 key bills on which the Senate voted during the 112th Congress, and which Democratic and Republican Senators voting “yea” and “nay” on each proposal.

Lastly, we obtained data on campaign contributions made by 92 PACs from the Open Secrets Website. We recorded all contributions made during the 2008, 2010, and 2012 election campaigns to the 102 persons who were Senators in the 112th Congress. The vast majority of PAC contributions to a candidate during a campaign was for $10,000 (the legal maximum is $5,000 each for a primary and the general election). We aggregated the contributions across all three electoral cycles, then dichotomized the sums into no contribution (0) and any contribution (1).

data(mpn_DemSxP)

data(mpn_RepSxP)

data(mpn_OverSxP)

Format

tbl_graph network object. It is a bimodal, directed, named, weighted graph of 51 Senators (type = FALSE) and 63 PACS (type = TRUE) and 2791 edges.

tbl_graph network object. It is a bimodal, directed, named, weighted graph of 62 Senators (type = FALSE) and 72 PACS (type = TRUE) and 3675 edges.

tbl_graph network object. It is a bimodal, directed, named, weighted graph of 20 Senators (type = FALSE) and 32 PACS (type = TRUE) and 614 edges.

Source

Knoke, Diani, Hollway, and Christopoulos. 2021. Multimodal Political Networks. Cambridge University Press: Cambridge.