Constraint for one- and two-mode networks

This function measures constraint for both one-mode and two-mode networks. For one-mode networks, the function wraps the implementation of Ron Burt's measure in {igraph}. For two-mode networks, the function employs the extension outlined in Hollway et al. (2020).

node_constraint(object, nodes = V(object), weights = NULL)

Arguments

object
A matrix, igraph graph, or tidygraph object.

nodes
The vertices for which the constraint will be calculated. Defaults to all vertices.

weights
The weights of the edges. If this is NULL and there is a weight edge attribute this is used. If there is no such edge attribute all edges will have the same weight.

Value

A named vector (one-mode) or a list of two named vectors ($nodes1, $nodes2).

References

Hollway, James, Jean-Frédéric Morin, and Joost Pauwelyn. 2020. "Structural conditions for novelty: the introduction of new environmental clauses to the trade regime complex." International Environmental Agreements: Politics, Law and Economics 20 (1): 61–83.

See also

Examples

node_constraint(southern_women)

#> $nodes1
#>    EVELYN     LAURA   THERESA    BRENDA CHARLOTTE   FRANCES   ELEANOR     PEARL 
#> 0.2762653 0.3064327 0.2939747 0.3165819 0.3836124 0.4124206 0.4153548 0.4610301 
#>      RUTH     VERNE      MYRA KATHERINE    SYLVIA      NORA     HELEN   DOROTHY 
#> 0.3980074 0.4070206 0.4340622 0.3458163 0.3227868 0.2575799 0.2776485 0.4340622 
#>    OLIVIA     FLORA 
#> 0.6109484 0.6109484 
#> 
#> $nodes2
#>        E1        E2        E3        E4        E5        E6        E7        E8 
#> 0.4763417 0.4755694 0.3380092 0.3938692 0.2745576 0.2405726 0.1924070 0.1703620 
#>        E9       E10       E11       E12       E13       E14 
#> 0.1605623 0.3266384 0.3078799 0.2920485 0.4317052 0.4317052 
#>