Watts-Strogatz small-world model for two-mode networks

Calculates small-world metrics for two-mode networks

node_smallworld(object, niter = 100)

Arguments

object

A matrix, igraph graph, or tidygraph object

niter

Number of simulations

Value

Returns a table of small-world related metrics for each second-mode node.

Details

The first column of the returned table is simply the number of the second-mode column. The next three columns report the observed and expected clustering, and the ratio of the former to the latter. The next three columns report the observed and expected path-length, and the ratio of the former to the later. The last column reports the ratio of the observed/expected clustering ratio to the observed/expected path-length ratio, which is known as a small-world metric. Expected clustering and paths is the mean of twomode_clustering and mean_distance over 100 random simulations with the same row and column sums.

See also

Examples

node_smallworld(southern_women)
#>    Num  ObsClust  ExpClust  ClustRat  ObsPath  ExpPath   PathRat SmallWorld
#> 1    1        NA        NA        NA       NA       NA        NA         NA
#> 2    2 0.5000000 0.4000000 1.2500000 1.866667 1.858667 1.0043042  1.2446429
#> 3    3 0.5600000 0.4720224 1.1863843 1.833333 2.224236 0.8242530  1.4393448
#> 4    4 0.5882353 0.5914268 0.9946037 1.822222 2.199310 0.8285426  1.2004256
#> 5    5 0.6268657 0.6566178 0.9546889 1.897436 2.306667 0.8225878  1.1605921
#> 6    6 0.6129032 0.6874713 0.8915328 2.033333 2.399333 0.8474576  1.0520087
#> 7    7 0.5544554 0.6427287 0.8626585 2.152632 2.489789 0.8645838  0.9977732
#> 8    8 0.5388740 0.6592821 0.8173648 2.108696 2.519130 0.8370728  0.9764560
#> 9    9 0.5095917 0.6146877 0.8290255 2.165242 2.480855 0.8727807  0.9498668
#> 10  10 0.4997785 0.5479366 0.9121101 2.195767 2.427037 0.9047110  1.0081784
#> 11  11 0.4854123 0.5039681 0.9631805 2.221675 2.406847 0.9230643  1.0434598
#> 12  12 0.4952520 0.4806701 1.0303367 2.239080 2.375540 0.9425563  1.0931301
#> 13  13 0.4888287 0.4736979 1.0319418 2.266667 2.395140 0.9463609  1.0904315
#> 14  14 0.4871634 0.4730360 1.0298654 2.290323 2.412097 0.9495152  1.0846223