Tools for reformatting networks, graphs, and matrices

Note that to_onemode(), which is currently only implemented for igraph, is not the same as project_rows() and project_cols(). There is no transformation involved; to_onemode() simply deletes the 'type' attribute from vertices, removing the bipartite note, but retaining all vertices.

to_unweighted(object, threshold = 1)

to_unnamed(object)

to_undirected(object)

to_onemode(object)

to_main_component(object)

to_uniplex(object, edge)

to_unsigned(object, keep = c("positive", "negative"))

to_simplex(object)

to_named(object)

to_multilevel(object)

Arguments

object
A matrix, {igraph} graph, {tidygraph} tbl_graph, or {network} object.

threshold
For a matrix, the threshold to binarise/dichotomise at.

edge
the name of an edge attribute to retain from a graph

keep
in the case of a signed network, whether to retain the "positive" or "negative" ties

Examples

to_unweighted(project_rows(southern_women))

#> IGRAPH 3a1a4b2 UN-- 18 139 -- 
#> + attr: name (v/c)
#> + edges from 3a1a4b2 (vertex names):
#>  [1] EVELYN --LAURA     EVELYN --BRENDA    EVELYN --THERESA   EVELYN --CHARLOTTE
#>  [5] EVELYN --FRANCES   EVELYN --ELEANOR   EVELYN --RUTH      EVELYN --PEARL    
#>  [9] EVELYN --NORA      EVELYN --VERNE     EVELYN --MYRA      EVELYN --KATHERINE
#> [13] EVELYN --SYLVIA    EVELYN --HELEN     EVELYN --DOROTHY   EVELYN --OLIVIA   
#> [17] EVELYN --FLORA     LAURA  --BRENDA    LAURA  --THERESA   LAURA  --CHARLOTTE
#> [21] LAURA  --FRANCES   LAURA  --ELEANOR   LAURA  --RUTH      LAURA  --PEARL    
#> [25] LAURA  --NORA      LAURA  --VERNE     LAURA  --SYLVIA    LAURA  --HELEN    
#> [29] LAURA  --MYRA      LAURA  --KATHERINE LAURA  --DOROTHY   THERESA--BRENDA   
#> + ... omitted several edges

to_unnamed(project_rows(southern_women))

#> IGRAPH eda2c29 U-W- 18 139 -- 
#> + attr: weight (e/n)
#> + edges from eda2c29:
#>   [1] 1-- 2 1-- 4 1-- 3 1-- 5 1-- 6 1-- 7 1-- 9 1-- 8 1--14 1--10 1--11 1--12
#>  [13] 1--13 1--15 1--16 1--17 1--18 2-- 4 2-- 3 2-- 5 2-- 6 2-- 7 2-- 9 2-- 8
#>  [25] 2--14 2--10 2--13 2--15 2--11 2--12 2--16 3-- 4 3-- 5 3-- 6 3-- 7 3-- 9
#>  [37] 3-- 8 3--14 3--10 3--13 3--15 3--11 3--12 3--16 3--17 3--18 4-- 5 4-- 6
#>  [49] 4-- 7 4-- 9 4-- 8 4--14 4--10 4--13 4--15 4--11 4--12 4--16 5-- 6 5-- 7
#>  [61] 5-- 9 5--10 5--13 5--14 5--15 6-- 7 6-- 9 6-- 8 6--14 6--10 6--11 6--12
#>  [73] 6--13 6--15 6--16 7-- 9 7-- 8 7--14 7--10 7--13 7--15 7--11 7--12 7--16
#>  [85] 8--14 8-- 9 8--10 8--11 8--12 8--13 8--15 8--16 8--17 8--18 9--10 9--13
#> + ... omitted several edges

to_undirected(ison_coleman)

#> # A tbl_graph: 8 nodes and 10 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 8 × 1 (active)
#>   name 
#>   <chr>
#> 1 Betty
#> 2 Sue  
#> 3 Alice
#> 4 Jane 
#> 5 Dale 
#> 6 Pam  
#> # … with 2 more rows
#> #
#> # Edge Data: 10 × 2
#>    from    to
#>   <int> <int>
#> 1     1     2
#> 2     2     3
#> 3     3     4
#> # … with 7 more rows

to_onemode(ison_marvel_teams)

#> IGRAPH 1f61f58 UN-- 194 683 -- 
#> + attr: name (v/c)
#> + edges from 1f61f58 (vertex names):
#>  [1] Abomination--Legion.Accursed         Abomination--The.Chronos.Corps      
#>  [3] Abomination--The.Strangers.Prisoners Abomination--The.Vi.Locks           
#>  [5] Abomination--Weapon.Plus             Ant-Man    --A.Next                 
#>  [7] Ant-Man    --Avengers                Ant-Man    --Avengers.Crew          
#>  [9] Ant-Man    --Defenders               Ant-Man    --Fantastic.Four         
#> [11] Ant-Man    --Heroes.For.Hire         Ant-Man    --Stark.Industries       
#> [13] Ant-Man    --Thunder.Guard           Ant-Man    --Ultimates              
#> [15] Apocalypse --Clan.Akkaba             Apocalypse --Horsemen.of.Apocalypse 
#> + ... omitted several edges

to_uniplex(ison_m182, "friend_tie")

#> # A tbl_graph: 16 nodes and 62 edges
#> #
#> # A directed simple graph with 3 components
#> #
#> # Node Data: 16 × 1 (active)
#>   name 
#>   <chr>
#> 1 1    
#> 2 2    
#> 3 3    
#> 4 4    
#> 5 5    
#> 6 6    
#> # … with 10 more rows
#> #
#> # Edge Data: 62 × 3
#>    from    to weight
#>   <int> <int>  <dbl>
#> 1     2     1      1
#> 2     2     7      1
#> 3     2     8      1
#> # … with 59 more rows

to_unsigned(ison_marvel_relationships, "positive")

#> # A tbl_graph: 53 nodes and 277 edges
#> #
#> # An undirected simple graph with 6 components
#> #
#> # Node Data: 53 × 10 (active)
#>   name  Gender Appearances Attractive  Rich Intellect Omnilingual PowerOrigin
#>   <chr> <chr>        <int>      <int> <int>     <int>       <int> <chr>      
#> 1 Abom… Male           427          0     0         1           1 Radiation  
#> 2 Ant-… Male           589          1     0         1           0 Human      
#> 3 Apoc… Male          1207          0     0         1           1 Mutant     
#> 4 Beast Male          7609          1     0         1           0 Mutant     
#> 5 Blac… Male          2189          1     1         1           0 Human      
#> 6 Blac… Female        2907          1     0         1           0 Human      
#> # … with 47 more rows, and 2 more variables: UnarmedCombat <int>,
#> #   ArmedCombat <int>
#> #
#> # Edge Data: 277 × 2
#>    from    to
#>   <int> <int>
#> 1     2    25
#> 2     2    29
#> 3     2    44
#> # … with 274 more rows

to_unsigned(ison_marvel_relationships, "negative")

#> # A tbl_graph: 53 nodes and 277 edges
#> #
#> # An undirected simple graph with 6 components
#> #
#> # Node Data: 53 × 10 (active)
#>   name  Gender Appearances Attractive  Rich Intellect Omnilingual PowerOrigin
#>   <chr> <chr>        <int>      <int> <int>     <int>       <int> <chr>      
#> 1 Abom… Male           427          0     0         1           1 Radiation  
#> 2 Ant-… Male           589          1     0         1           0 Human      
#> 3 Apoc… Male          1207          0     0         1           1 Mutant     
#> 4 Beast Male          7609          1     0         1           0 Mutant     
#> 5 Blac… Male          2189          1     1         1           0 Human      
#> 6 Blac… Female        2907          1     0         1           0 Human      
#> # … with 47 more rows, and 2 more variables: UnarmedCombat <int>,
#> #   ArmedCombat <int>
#> #
#> # Edge Data: 277 × 2
#>    from    to
#>   <int> <int>
#> 1     2    25
#> 2     2    29
#> 3     2    44
#> # … with 274 more rows

to_simplex(ison_m182)

#> # A tbl_graph: 16 nodes and 144 edges
#> #
#> # A directed simple graph with 1 component
#> #
#> # Node Data: 16 × 1 (active)
#>   name 
#>   <chr>
#> 1 1    
#> 2 2    
#> 3 3    
#> 4 4    
#> 5 5    
#> 6 6    
#> # … with 10 more rows
#> #
#> # Edge Data: 144 × 2
#>    from    to
#>   <int> <int>
#> 1     1     5
#> 2     1     8
#> 3     1     9
#> # … with 141 more rows

to_named(ison_m182)

#> # A tbl_graph: 16 nodes and 144 edges
#> #
#> # A directed simple graph with 1 component
#> #
#> # Node Data: 16 × 1 (active)
#>   name  
#>   <chr> 
#> 1 Dwight
#> 2 Jennie
#> 3 Karla 
#> 4 Caleb 
#> 5 Aria  
#> 6 Brody 
#> # … with 10 more rows
#> #
#> # Edge Data: 144 × 5
#>    from    to friend_tie social_tie task_tie
#>   <int> <int>      <dbl>      <dbl>    <dbl>
#> 1     1     5          0       1.2       0.3
#> 2     1     8          0       0.15      0  
#> 3     1     9          0       2.85      0.3
#> # … with 141 more rows

to_multilevel(mpn_elite_usa_advice)

#> # A tbl_graph: 34 nodes and 46 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 34 × 2 (active)
#>   name       lvl
#>   <chr>    <dbl>
#> 1 Albright     1
#> 2 Argyros      1
#> 3 Armitage     1
#> 4 Curry        1
#> 5 Fukuyama     1
#> 6 Gray         1
#> # … with 28 more rows
#> #
#> # Edge Data: 46 × 2
#>    from    to
#>   <int> <int>
#> 1     1    17
#> 2     1    19
#> 3     1    21
#> # … with 43 more rows